chore(gpu): split integer utilities file
