+++ backends/tfhe-cuda-backend/cuda/include/integer/integer_utilities.h
@@ -907,7 +907,6 @@ struct int_noise_squashing_lut
 
 // Forward declarations for operation buffers
 template <typename Torus> struct int_sub_and_propagate;
-template <typename Torus> struct int_bitop_buffer;
 
 template <typename Torus> struct int_bit_extract_luts_buffer {
   int_radix_params params;
@@ -1007,173 +1006,6 @@ template <typename Torus> struct int_bit_extract_luts_buffer {
   }
 };
 
-template <typename Torus> struct int_shift_and_rotate_buffer {
-  int_radix_params params;
-  SHIFT_OR_ROTATE_TYPE shift_type;
-  bool is_signed;
-
-  CudaRadixCiphertextFFI *tmp_bits;
-  CudaRadixCiphertextFFI *tmp_shift_bits;
-  CudaRadixCiphertextFFI *tmp_rotated;
-  CudaRadixCiphertextFFI *tmp_input_bits_a;
-  CudaRadixCiphertextFFI *tmp_input_bits_b;
-  CudaRadixCiphertextFFI *tmp_mux_inputs;
-
-  int_bit_extract_luts_buffer<Torus> *bit_extract_luts;
-  int_bit_extract_luts_buffer<Torus> *bit_extract_luts_with_offset_2;
-  int_radix_lut<Torus> *mux_lut;
-  int_radix_lut<Torus> *cleaning_lut;
-
-  Torus offset;
-  bool gpu_memory_allocated;
-
-  int_shift_and_rotate_buffer(CudaStreams streams,
-                              SHIFT_OR_ROTATE_TYPE shift_type, bool is_signed,
-                              int_radix_params params,
-                              uint32_t num_radix_blocks,
-                              bool allocate_gpu_memory,
-                              uint64_t &size_tracker) {
-    this->shift_type = shift_type;
-    this->is_signed = is_signed;
-    this->params = params;
-    gpu_memory_allocated = allocate_gpu_memory;
-
-    uint32_t bits_per_block = std::log2(params.message_modulus);
-    uint32_t total_nb_bits =
-        std::log2(params.message_modulus) * num_radix_blocks;
-    uint32_t max_num_bits_that_tell_shift = std::log2(total_nb_bits);
-
-    auto is_power_of_two = [](uint32_t n) {
-      return (n > 0) && ((n & (n - 1)) == 0);
-    };
-
-    if (!is_power_of_two(total_nb_bits))
-      max_num_bits_that_tell_shift += 1;
-
-    offset = (shift_type == LEFT_SHIFT ? 0 : total_nb_bits);
-
-    bit_extract_luts = new int_bit_extract_luts_buffer<Torus>(
-        streams, params, bits_per_block, num_radix_blocks, allocate_gpu_memory,
-        size_tracker);
-    bit_extract_luts_with_offset_2 = new int_bit_extract_luts_buffer<Torus>(
-        streams, params, bits_per_block, 2, num_radix_blocks,
-        allocate_gpu_memory, size_tracker);
-
-    mux_lut = new int_radix_lut<Torus>(streams, params, 1,
-                                       bits_per_block * num_radix_blocks,
-                                       allocate_gpu_memory, size_tracker);
-    cleaning_lut =
-        new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                 allocate_gpu_memory, size_tracker);
-
-    tmp_bits = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_bits,
-        bits_per_block * num_radix_blocks, params.big_lwe_dimension,
-        size_tracker, allocate_gpu_memory);
-
-    tmp_shift_bits = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_shift_bits,
-        max_num_bits_that_tell_shift * num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-    tmp_rotated = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_rotated,
-        bits_per_block * num_radix_blocks, params.big_lwe_dimension,
-        size_tracker, allocate_gpu_memory);
-
-    tmp_input_bits_a = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_input_bits_a,
-        bits_per_block * num_radix_blocks, params.big_lwe_dimension,
-        size_tracker, allocate_gpu_memory);
-
-    tmp_input_bits_b = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_input_bits_b,
-        bits_per_block * num_radix_blocks, params.big_lwe_dimension,
-        size_tracker, allocate_gpu_memory);
-
-    tmp_mux_inputs = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_mux_inputs,
-        bits_per_block * num_radix_blocks, params.big_lwe_dimension,
-        size_tracker, allocate_gpu_memory);
-
-    auto mux_lut_f = [](Torus x) -> Torus {
-      // x is expected to be x = 0bcba
-      // where
-      // - c is the control bit
-      // - b the bit value returned if c is 1
-      // - a the bit value returned if c is 0
-      // (any bit above c is ignored)
-      x = x & 7;
-      auto control_bit = x >> 2;
-      auto previous_bit = (x & 2) >> 1;
-      auto current_bit = x & 1;
-
-      if (control_bit == 1)
-        return previous_bit;
-      else
-        return current_bit;
-    };
-
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0), mux_lut->get_lut(0, 0),
-        mux_lut->get_degree(0), mux_lut->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, mux_lut_f, gpu_memory_allocated);
-    auto active_gpu_count_mux =
-        streams.active_gpu_subset(bits_per_block * num_radix_blocks);
-    mux_lut->broadcast_lut(active_gpu_count_mux);
-
-    auto cleaning_lut_f = [params](Torus x) -> Torus {
-      return x % params.message_modulus;
-    };
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0), cleaning_lut->get_lut(0, 0),
-        cleaning_lut->get_degree(0), cleaning_lut->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, cleaning_lut_f, gpu_memory_allocated);
-    auto active_gpu_count_cleaning =
-        streams.active_gpu_subset(num_radix_blocks);
-    cleaning_lut->broadcast_lut(active_gpu_count_cleaning);
-  }
-
-  void release(CudaStreams streams) {
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_bits, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_shift_bits, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_rotated, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_input_bits_a, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_input_bits_b, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_mux_inputs, gpu_memory_allocated);
-
-    bit_extract_luts->release(streams);
-    bit_extract_luts_with_offset_2->release(streams);
-    mux_lut->release(streams);
-    cleaning_lut->release(streams);
-
-    delete tmp_bits;
-    delete tmp_shift_bits;
-    delete tmp_rotated;
-    delete tmp_input_bits_a;
-    delete tmp_input_bits_b;
-    delete tmp_mux_inputs;
-    delete bit_extract_luts;
-    delete bit_extract_luts_with_offset_2;
-    delete mux_lut;
-    delete cleaning_lut;
-  }
-};
-
 template <typename Torus> struct int_fullprop_buffer {
   int_radix_params params;
 
@@ -1251,137 +1083,6 @@ template <typename Torus> struct int_fullprop_buffer {
     delete tmp_small_lwe_vector;
     delete tmp_big_lwe_vector;
     delete lut;
-  }
-};
-
-template <typename Torus> struct int_overflowing_sub_memory {
-  Torus *generates_or_propagates;
-  Torus *step_output;
-
-  int_radix_lut<Torus> *luts_array;
-  int_radix_lut<Torus> *luts_borrow_propagation_sum;
-  int_radix_lut<Torus> *message_acc;
-
-  int_radix_params params;
-  bool gpu_memory_allocated;
-
-  int_overflowing_sub_memory(CudaStreams streams, int_radix_params params,
-                             uint32_t num_radix_blocks,
-                             bool allocate_gpu_memory, uint64_t &size_tracker) {
-    this->params = params;
-    gpu_memory_allocated = allocate_gpu_memory;
-    auto glwe_dimension = params.glwe_dimension;
-    auto polynomial_size = params.polynomial_size;
-    auto message_modulus = params.message_modulus;
-    auto carry_modulus = params.carry_modulus;
-    auto big_lwe_size = (polynomial_size * glwe_dimension + 1);
-    auto big_lwe_size_bytes = big_lwe_size * sizeof(Torus);
-
-    // allocate memory for intermediate calculations
-    generates_or_propagates = (Torus *)cuda_malloc_with_size_tracking_async(
-        num_radix_blocks * big_lwe_size_bytes, streams.stream(0),
-        streams.gpu_index(0), size_tracker, allocate_gpu_memory);
-    step_output = (Torus *)cuda_malloc_with_size_tracking_async(
-        num_radix_blocks * big_lwe_size_bytes, streams.stream(0),
-        streams.gpu_index(0), size_tracker, allocate_gpu_memory);
-    cuda_memset_with_size_tracking_async(
-        generates_or_propagates, 0, num_radix_blocks * big_lwe_size_bytes,
-        streams.stream(0), streams.gpu_index(0), allocate_gpu_memory);
-    cuda_memset_with_size_tracking_async(
-        step_output, 0, num_radix_blocks * big_lwe_size_bytes,
-        streams.stream(0), streams.gpu_index(0), allocate_gpu_memory);
-
-    // declare functions for lut generation
-    auto f_lut_does_block_generate_carry = [message_modulus](Torus x) -> Torus {
-      if (x < message_modulus)
-        return OUTPUT_CARRY::GENERATED;
-      return OUTPUT_CARRY::NONE;
-    };
-
-    auto f_lut_does_block_generate_or_propagate =
-        [message_modulus](Torus x) -> Torus {
-      if (x < message_modulus)
-        return OUTPUT_CARRY::GENERATED;
-      else if (x == message_modulus)
-        return OUTPUT_CARRY::PROPAGATED;
-      return OUTPUT_CARRY::NONE;
-    };
-
-    auto f_luts_borrow_propagation_sum = [](Torus msb, Torus lsb) -> Torus {
-      if (msb == OUTPUT_CARRY::PROPAGATED)
-        return lsb;
-      return msb;
-    };
-
-    auto f_message_acc = [message_modulus](Torus x) -> Torus {
-      return x % message_modulus;
-    };
-
-    // create lut objects
-    luts_array = new int_radix_lut<Torus>(streams, params, 2, num_radix_blocks,
-                                          allocate_gpu_memory, size_tracker);
-    luts_borrow_propagation_sum = new int_radix_lut<Torus>(
-        streams, params, 1, num_radix_blocks, luts_array, size_tracker,
-        allocate_gpu_memory, size_tracker);
-    message_acc = new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                           luts_array, size_tracker,
-                                           allocate_gpu_memory, size_tracker);
-
-    auto lut_does_block_generate_carry = luts_array->get_lut(0, 0);
-    auto lut_does_block_generate_or_propagate = luts_array->get_lut(0, 1);
-
-    // generate luts (aka accumulators)
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0), lut_does_block_generate_carry,
-        luts_array->get_degree(0), luts_array->get_max_degree(0),
-        glwe_dimension, polynomial_size, message_modulus, carry_modulus,
-        f_lut_does_block_generate_carry, gpu_memory_allocated);
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0),
-        lut_does_block_generate_or_propagate, luts_array->get_degree(1),
-        luts_array->get_max_degree(1), glwe_dimension, polynomial_size,
-        message_modulus, carry_modulus, f_lut_does_block_generate_or_propagate,
-        gpu_memory_allocated);
-    if (allocate_gpu_memory)
-      cuda_set_value_async<Torus>(streams.stream(0), streams.gpu_index(0),
-                                  luts_array->get_lut_indexes(0, 1), 1,
-                                  num_radix_blocks - 1);
-
-    generate_device_accumulator_bivariate<Torus>(
-        streams.stream(0), streams.gpu_index(0),
-        luts_borrow_propagation_sum->get_lut(0, 0),
-        luts_borrow_propagation_sum->get_degree(0),
-        luts_borrow_propagation_sum->get_max_degree(0), glwe_dimension,
-        polynomial_size, message_modulus, carry_modulus,
-        f_luts_borrow_propagation_sum, gpu_memory_allocated);
-
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0), message_acc->get_lut(0, 0),
-        message_acc->get_degree(0), message_acc->get_max_degree(0),
-        glwe_dimension, polynomial_size, message_modulus, carry_modulus,
-        f_message_acc, gpu_memory_allocated);
-
-    auto active_streams = streams.active_gpu_subset(num_radix_blocks);
-    luts_array->broadcast_lut(active_streams);
-    luts_borrow_propagation_sum->broadcast_lut(active_streams);
-    message_acc->broadcast_lut(active_streams);
-  }
-
-  void release(CudaStreams streams) {
-    cuda_drop_with_size_tracking_async(generates_or_propagates,
-                                       streams.stream(0), streams.gpu_index(0),
-                                       gpu_memory_allocated);
-    cuda_drop_with_size_tracking_async(step_output, streams.stream(0),
-                                       streams.gpu_index(0),
-                                       gpu_memory_allocated);
-
-    luts_array->release(streams);
-    luts_borrow_propagation_sum->release(streams);
-    message_acc->release(streams);
-
-    delete luts_array;
-    delete luts_borrow_propagation_sum;
-    delete message_acc;
   }
 };
 
@@ -2828,4179 +2529,6 @@ template <typename Torus> struct int_borrow_prop_memory {
     sub_streams_2.release();
   };
 };
-
-template <typename Torus> struct int_zero_out_if_buffer {
-
-  int_radix_params params;
-
-  CudaRadixCiphertextFFI *tmp;
-
-  bool gpu_memory_allocated;
-
-  int_zero_out_if_buffer(CudaStreams streams, int_radix_params params,
-                         uint32_t num_radix_blocks, bool allocate_gpu_memory,
-                         uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-    this->params = params;
-    auto active_streams = streams.active_gpu_subset(num_radix_blocks);
-
-    tmp = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp, num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-  }
-  void release(CudaStreams streams) {
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0), tmp,
-                                   gpu_memory_allocated);
-    delete tmp;
-    tmp = nullptr;
-  }
-};
-
-template <typename Torus> struct int_mul_memory {
-  CudaRadixCiphertextFFI *vector_result_sb;
-  CudaRadixCiphertextFFI *block_mul_res;
-  CudaRadixCiphertextFFI *small_lwe_vector;
-
-  int_radix_lut<Torus> *luts_array; // lsb msb
-  int_radix_lut<Torus> *zero_out_predicate_lut;
-
-  int_sum_ciphertexts_vec_memory<Torus> *sum_ciphertexts_mem;
-  int_sc_prop_memory<Torus> *sc_prop_mem;
-  int_zero_out_if_buffer<Torus> *zero_out_mem;
-
-  int_radix_params params;
-  bool boolean_mul = false;
-  bool gpu_memory_allocated;
-
-  int_mul_memory(CudaStreams streams, int_radix_params params,
-                 bool const is_boolean_left, bool const is_boolean_right,
-                 uint32_t num_radix_blocks, bool allocate_gpu_memory,
-                 uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-    this->boolean_mul = is_boolean_left || is_boolean_right;
-    this->params = params;
-
-    if (boolean_mul) {
-      auto zero_out_predicate_lut_f = [](Torus block,
-                                         Torus condition) -> Torus {
-        if (condition == 0)
-          return 0;
-        else
-          return block;
-      };
-      zero_out_predicate_lut =
-          new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                   allocate_gpu_memory, size_tracker);
-      generate_device_accumulator_bivariate<Torus>(
-          streams.stream(0), streams.gpu_index(0),
-          zero_out_predicate_lut->get_lut(0, 0),
-          zero_out_predicate_lut->get_degree(0),
-          zero_out_predicate_lut->get_max_degree(0), params.glwe_dimension,
-          params.polynomial_size, params.message_modulus, params.carry_modulus,
-          zero_out_predicate_lut_f, gpu_memory_allocated);
-
-      auto active_streams = streams.active_gpu_subset(num_radix_blocks);
-      zero_out_predicate_lut->broadcast_lut(active_streams);
-
-      zero_out_mem = new int_zero_out_if_buffer<Torus>(
-          streams, params, num_radix_blocks, allocate_gpu_memory, size_tracker);
-
-      return;
-    }
-
-    auto glwe_dimension = params.glwe_dimension;
-    auto polynomial_size = params.polynomial_size;
-    auto message_modulus = params.message_modulus;
-    auto carry_modulus = params.carry_modulus;
-
-    // 'vector_result_lsb' contains blocks from all possible shifts of
-    // radix_lwe_left excluding zero ciphertext blocks
-    int lsb_vector_block_count = num_radix_blocks * (num_radix_blocks + 1) / 2;
-
-    // 'vector_result_msb' contains blocks from all possible shifts of
-    // radix_lwe_left except the last blocks of each shift
-    int msb_vector_block_count = num_radix_blocks * (num_radix_blocks - 1) / 2;
-
-    int total_block_count = num_radix_blocks * num_radix_blocks;
-
-    // allocate memory for intermediate buffers
-    vector_result_sb = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), vector_result_sb,
-        2 * total_block_count, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-    block_mul_res = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), block_mul_res,
-        2 * total_block_count, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-    small_lwe_vector = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), small_lwe_vector,
-        2 * total_block_count, params.small_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    // create int_radix_lut objects for lsb, msb, message, carry
-    // luts_array -> lut = {lsb_acc, msb_acc}
-    luts_array = new int_radix_lut<Torus>(streams, params, 2, total_block_count,
-                                          allocate_gpu_memory, size_tracker);
-    auto lsb_acc = luts_array->get_lut(0, 0);
-    auto msb_acc = luts_array->get_lut(0, 1);
-
-    // define functions for each accumulator
-    auto lut_f_lsb = [message_modulus](Torus x, Torus y) -> Torus {
-      return (x * y) % message_modulus;
-    };
-    auto lut_f_msb = [message_modulus](Torus x, Torus y) -> Torus {
-      return (x * y) / message_modulus;
-    };
-
-    // generate accumulators
-    generate_device_accumulator_bivariate<Torus>(
-        streams.stream(0), streams.gpu_index(0), lsb_acc,
-        luts_array->get_degree(0), luts_array->get_max_degree(0),
-        glwe_dimension, polynomial_size, message_modulus, carry_modulus,
-        lut_f_lsb, gpu_memory_allocated);
-    generate_device_accumulator_bivariate<Torus>(
-        streams.stream(0), streams.gpu_index(0), msb_acc,
-        luts_array->get_degree(1), luts_array->get_max_degree(1),
-        glwe_dimension, polynomial_size, message_modulus, carry_modulus,
-        lut_f_msb, gpu_memory_allocated);
-
-    // lut_indexes_vec for luts_array should be reinitialized
-    // first lsb_vector_block_count value should reference to lsb_acc
-    // last msb_vector_block_count values should reference to msb_acc
-    // for message and carry default lut_indexes_vec is fine
-    if (allocate_gpu_memory)
-      cuda_set_value_async<Torus>(
-          streams.stream(0), streams.gpu_index(0),
-          luts_array->get_lut_indexes(0, lsb_vector_block_count), 1,
-          msb_vector_block_count);
-    auto active_streams = streams.active_gpu_subset(total_block_count);
-    luts_array->broadcast_lut(active_streams);
-    // create memory object for sum ciphertexts
-    sum_ciphertexts_mem = new int_sum_ciphertexts_vec_memory<Torus>(
-        streams, params, num_radix_blocks, 2 * num_radix_blocks,
-        vector_result_sb, small_lwe_vector, luts_array, true,
-        allocate_gpu_memory, size_tracker);
-    uint32_t requested_flag = outputFlag::FLAG_NONE;
-    sc_prop_mem = new int_sc_prop_memory<Torus>(
-        streams, params, num_radix_blocks, requested_flag, allocate_gpu_memory,
-        size_tracker);
-  }
-
-  void release(CudaStreams streams) {
-
-    if (boolean_mul) {
-      zero_out_predicate_lut->release(streams);
-      zero_out_mem->release(streams);
-      delete zero_out_mem;
-      delete zero_out_predicate_lut;
-
-      return;
-    }
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   vector_result_sb, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   block_mul_res, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   small_lwe_vector, gpu_memory_allocated);
-
-    luts_array->release(streams);
-    sum_ciphertexts_mem->release(streams);
-    sc_prop_mem->release(streams);
-
-    delete vector_result_sb;
-    delete block_mul_res;
-    delete small_lwe_vector;
-    delete luts_array;
-    delete sum_ciphertexts_mem;
-    delete sc_prop_mem;
-  }
-};
-
-template <typename Torus> struct int_logical_scalar_shift_buffer {
-  int_radix_params params;
-  std::vector<int_radix_lut<Torus> *> lut_buffers_bivariate;
-
-  SHIFT_OR_ROTATE_TYPE shift_type;
-
-  CudaRadixCiphertextFFI *tmp_rotated;
-
-  bool reuse_memory = false;
-  bool gpu_memory_allocated;
-
-  int_logical_scalar_shift_buffer(CudaStreams streams,
-                                  SHIFT_OR_ROTATE_TYPE shift_type,
-                                  int_radix_params params,
-                                  uint32_t num_radix_blocks,
-                                  bool allocate_gpu_memory,
-                                  uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-    this->shift_type = shift_type;
-    this->params = params;
-
-    uint32_t max_amount_of_pbs = num_radix_blocks;
-
-    tmp_rotated = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_rotated,
-        max_amount_of_pbs + 2, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    uint32_t num_bits_in_block = (uint32_t)std::log2(params.message_modulus);
-
-    // LUT
-    // pregenerate lut vector and indexes
-    // lut for left shift
-    // here we generate 'num_bits_in_block' times lut
-    // one for each 'shift_within_block' = 'shift' % 'num_bits_in_block'
-    // even though lut_left contains 'num_bits_in_block' lut
-    // lut_indexes_vec will have indexes for single lut only and those indexes
-    // will be 0 it means for pbs corresponding lut should be selected and
-    // pass along lut_indexes_vec filled with zeros
-
-    // calculate bivariate lut for each 'shift_within_block'
-    // so that in case an application calls scratches only once for a whole
-    // circuit it can reuse memory for different shift values
-    for (int s_w_b = 1; s_w_b < num_bits_in_block; s_w_b++) {
-      auto cur_lut_bivariate =
-          new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                   allocate_gpu_memory, size_tracker);
-
-      uint32_t shift_within_block = s_w_b;
-
-      std::function<Torus(Torus, Torus)> shift_lut_f;
-
-      if (shift_type == LEFT_SHIFT) {
-        shift_lut_f = [shift_within_block, params](
-                          Torus current_block, Torus previous_block) -> Torus {
-          current_block = current_block << shift_within_block;
-          previous_block = previous_block << shift_within_block;
-
-          Torus message_of_current_block =
-              current_block % params.message_modulus;
-          Torus carry_of_previous_block =
-              previous_block / params.message_modulus;
-          return message_of_current_block + carry_of_previous_block;
-        };
-      } else {
-        shift_lut_f = [num_bits_in_block, shift_within_block,
-                       params](Torus current_block, Torus next_block) -> Torus {
-          // left shift so as not to lose
-          // bits when shifting right afterwards
-          next_block <<= num_bits_in_block;
-          next_block >>= shift_within_block;
-
-          // The way of getting carry / message is reversed compared
-          // to the usual way but its normal:
-          // The message is in the upper bits, the carry in lower bits
-          Torus message_of_current_block = current_block >> shift_within_block;
-          Torus carry_of_previous_block = next_block % params.message_modulus;
-
-          return message_of_current_block + carry_of_previous_block;
-        };
-      }
-
-      // right shift
-      generate_device_accumulator_bivariate<Torus>(
-          streams.stream(0), streams.gpu_index(0),
-          cur_lut_bivariate->get_lut(0, 0), cur_lut_bivariate->get_degree(0),
-          cur_lut_bivariate->get_max_degree(0), params.glwe_dimension,
-          params.polynomial_size, params.message_modulus, params.carry_modulus,
-          shift_lut_f, gpu_memory_allocated);
-      auto active_streams = streams.active_gpu_subset(num_radix_blocks);
-      cur_lut_bivariate->broadcast_lut(active_streams);
-
-      lut_buffers_bivariate.push_back(cur_lut_bivariate);
-    }
-  }
-
-  int_logical_scalar_shift_buffer(CudaStreams streams,
-                                  SHIFT_OR_ROTATE_TYPE shift_type,
-                                  int_radix_params params,
-                                  uint32_t num_radix_blocks,
-                                  bool allocate_gpu_memory,
-                                  CudaRadixCiphertextFFI *pre_allocated_buffer,
-                                  uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-    this->shift_type = shift_type;
-    this->params = params;
-    tmp_rotated = pre_allocated_buffer;
-    reuse_memory = true;
-
-    if (allocate_gpu_memory)
-      set_zero_radix_ciphertext_slice_async<Torus>(
-          streams.stream(0), streams.gpu_index(0), tmp_rotated, 0,
-          tmp_rotated->num_radix_blocks);
-
-    uint32_t num_bits_in_block = (uint32_t)std::log2(params.message_modulus);
-
-    // LUT
-    // pregenerate lut vector and indexes
-    // lut for left shift
-    // here we generate 'num_bits_in_block' times lut
-    // one for each 'shift_within_block' = 'shift' % 'num_bits_in_block'
-    // even though lut_left contains 'num_bits_in_block' lut
-    // lut_indexes_vec will have indexes for single lut only and those indexes
-    // will be 0 it means for pbs corresponding lut should be selected and
-    // pass along lut_indexes_vec filled with zeros
-
-    // calculate bivariate lut for each 'shift_within_block'
-    // so that in case an application calls scratches only once for a whole
-    // circuit it can reuse memory for different shift values
-    for (int s_w_b = 1; s_w_b < num_bits_in_block; s_w_b++) {
-      auto cur_lut_bivariate =
-          new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                   allocate_gpu_memory, size_tracker);
-
-      uint32_t shift_within_block = s_w_b;
-
-      std::function<Torus(Torus, Torus)> shift_lut_f;
-
-      if (shift_type == LEFT_SHIFT) {
-        shift_lut_f = [shift_within_block, params](
-                          Torus current_block, Torus previous_block) -> Torus {
-          current_block = current_block << shift_within_block;
-          previous_block = previous_block << shift_within_block;
-
-          Torus message_of_current_block =
-              current_block % params.message_modulus;
-          Torus carry_of_previous_block =
-              previous_block / params.message_modulus;
-          return message_of_current_block + carry_of_previous_block;
-        };
-      } else {
-        shift_lut_f = [num_bits_in_block, shift_within_block,
-                       params](Torus current_block, Torus next_block) -> Torus {
-          // left shift so as not to lose
-          // bits when shifting right afterwards
-          next_block <<= num_bits_in_block;
-          next_block >>= shift_within_block;
-
-          // The way of getting carry / message is reversed compared
-          // to the usual way but its normal:
-          // The message is in the upper bits, the carry in lower bits
-          Torus message_of_current_block = current_block >> shift_within_block;
-          Torus carry_of_previous_block = next_block % params.message_modulus;
-
-          return message_of_current_block + carry_of_previous_block;
-        };
-      }
-
-      // right shift
-      generate_device_accumulator_bivariate<Torus>(
-          streams.stream(0), streams.gpu_index(0),
-          cur_lut_bivariate->get_lut(0, 0), cur_lut_bivariate->get_degree(0),
-          cur_lut_bivariate->get_max_degree(0), params.glwe_dimension,
-          params.polynomial_size, params.message_modulus, params.carry_modulus,
-          shift_lut_f, gpu_memory_allocated);
-      auto active_streams = streams.active_gpu_subset(num_radix_blocks);
-      cur_lut_bivariate->broadcast_lut(active_streams);
-
-      lut_buffers_bivariate.push_back(cur_lut_bivariate);
-    }
-  }
-  void release(CudaStreams streams) {
-    cuda_synchronize_stream(streams.stream(0), streams.gpu_index(0));
-    for (auto &buffer : lut_buffers_bivariate) {
-      buffer->release(streams);
-      delete buffer;
-    }
-    lut_buffers_bivariate.clear();
-
-    if (!reuse_memory) {
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     tmp_rotated, gpu_memory_allocated);
-      cuda_synchronize_stream(streams.stream(0), streams.gpu_index(0));
-      delete tmp_rotated;
-    }
-  }
-};
-
-template <typename Torus> struct int_arithmetic_scalar_shift_buffer {
-  int_radix_params params;
-  std::vector<int_radix_lut<Torus> *> lut_buffers_univariate;
-  std::vector<int_radix_lut<Torus> *> lut_buffers_bivariate;
-
-  SHIFT_OR_ROTATE_TYPE shift_type;
-
-  CudaRadixCiphertextFFI *tmp_rotated;
-
-  CudaStreams local_streams_1;
-  CudaStreams local_streams_2;
-  bool gpu_memory_allocated;
-
-  int_arithmetic_scalar_shift_buffer(CudaStreams streams,
-                                     SHIFT_OR_ROTATE_TYPE shift_type,
-                                     int_radix_params params,
-                                     uint32_t num_radix_blocks,
-                                     bool allocate_gpu_memory,
-                                     uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-
-    auto active_streams = streams.active_gpu_subset(1);
-    // In the arithmetic shift, a PBS has to be applied to the last rotated
-    // block twice: once to shift it, once to compute the padding block to be
-    // copied onto all blocks to the left of the last rotated block
-    local_streams_1.create_on_same_gpus(active_streams);
-    local_streams_2.create_on_same_gpus(active_streams);
-    this->shift_type = shift_type;
-    this->params = params;
-
-    tmp_rotated = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_rotated,
-        num_radix_blocks + 3, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    uint32_t num_bits_in_block = (uint32_t)std::log2(params.message_modulus);
-
-    // LUT
-    // pregenerate lut vector and indexes lut
-
-    // lut to shift the last block
-    // calculate lut for each 'shift_within_block'
-    // so that in case an application calls scratches only once for a whole
-    // circuit it can reuse memory for different shift values
-    // With two bits of message this is actually only one LUT.
-    for (int s_w_b = 1; s_w_b < num_bits_in_block; s_w_b++) {
-      auto shift_last_block_lut_univariate = new int_radix_lut<Torus>(
-          streams, params, 1, 1, allocate_gpu_memory, size_tracker);
-
-      uint32_t shift_within_block = s_w_b;
-
-      std::function<Torus(Torus)> last_block_lut_f;
-      last_block_lut_f = [num_bits_in_block, shift_within_block,
-                          params](Torus x) -> Torus {
-        x = x % params.message_modulus;
-        uint32_t x_sign_bit = x >> (num_bits_in_block - 1) & 1;
-        uint32_t shifted = x >> shift_within_block;
-        // padding is a message full of 1 if sign bit is one
-        // else padding is a zero message
-        uint32_t padding = (params.message_modulus - 1) * x_sign_bit;
-
-        // Make padding have 1s only in places where bits
-        // where actually need to be padded
-        padding <<= num_bits_in_block - shift_within_block;
-        padding %= params.message_modulus;
-
-        return shifted | padding;
-      };
-
-      generate_device_accumulator<Torus>(
-          streams.stream(0), streams.gpu_index(0),
-          shift_last_block_lut_univariate->get_lut(0, 0),
-          shift_last_block_lut_univariate->get_degree(0),
-          shift_last_block_lut_univariate->get_max_degree(0),
-          params.glwe_dimension, params.polynomial_size, params.message_modulus,
-          params.carry_modulus, last_block_lut_f, gpu_memory_allocated);
-      auto active_streams_shift_last = streams.active_gpu_subset(1);
-      shift_last_block_lut_univariate->broadcast_lut(active_streams_shift_last);
-
-      lut_buffers_univariate.push_back(shift_last_block_lut_univariate);
-    }
-
-    auto padding_block_lut_univariate = new int_radix_lut<Torus>(
-        streams, params, 1, 1, allocate_gpu_memory, size_tracker);
-
-    // lut to compute the padding block
-    std::function<Torus(Torus)> padding_block_lut_f;
-    padding_block_lut_f = [num_bits_in_block, params](Torus x) -> Torus {
-      x = x % params.message_modulus;
-      uint32_t x_sign_bit = x >> (num_bits_in_block - 1) & 1;
-      // padding is a message full of 1 if sign bit is one
-      // else padding is a zero message
-      return (params.message_modulus - 1) * x_sign_bit;
-    };
-
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0),
-        padding_block_lut_univariate->get_lut(0, 0),
-        padding_block_lut_univariate->get_degree(0),
-        padding_block_lut_univariate->get_max_degree(0), params.glwe_dimension,
-        params.polynomial_size, params.message_modulus, params.carry_modulus,
-        padding_block_lut_f, gpu_memory_allocated);
-    // auto active_streams = streams.active_gpu_subset(1);
-    padding_block_lut_univariate->broadcast_lut(active_streams);
-
-    lut_buffers_univariate.push_back(padding_block_lut_univariate);
-
-    // lut to shift the first blocks
-    // calculate lut for each 'shift_within_block'
-    // so that in case an application calls scratches only once for a whole
-    // circuit it can reuse memory for different shift values
-    // NB: with two bits of message, this is actually only one LUT.
-    for (int s_w_b = 1; s_w_b < num_bits_in_block; s_w_b++) {
-      auto shift_blocks_lut_bivariate =
-          new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                   allocate_gpu_memory, size_tracker);
-
-      uint32_t shift_within_block = s_w_b;
-
-      std::function<Torus(Torus, Torus)> blocks_lut_f;
-      blocks_lut_f = [num_bits_in_block, shift_within_block,
-                      params](Torus current_block, Torus next_block) -> Torus {
-        // left shift so as not to lose
-        // bits when shifting right after
-        next_block <<= num_bits_in_block;
-        next_block >>= shift_within_block;
-
-        // The way of getting carry / message is reversed compared
-        // to the usual way but its normal:
-        // The message is in the upper bits, the carry in lower bits
-        uint32_t message_of_current_block = current_block >> shift_within_block;
-        uint32_t carry_of_previous_block = next_block % params.message_modulus;
-
-        return message_of_current_block + carry_of_previous_block;
-      };
-
-      generate_device_accumulator_bivariate<Torus>(
-          streams.stream(0), streams.gpu_index(0),
-          shift_blocks_lut_bivariate->get_lut(0, 0),
-          shift_blocks_lut_bivariate->get_degree(0),
-          shift_blocks_lut_bivariate->get_max_degree(0), params.glwe_dimension,
-          params.polynomial_size, params.message_modulus, params.carry_modulus,
-          blocks_lut_f, gpu_memory_allocated);
-      auto active_streams_shift_blocks =
-          streams.active_gpu_subset(num_radix_blocks);
-      shift_blocks_lut_bivariate->broadcast_lut(active_streams_shift_blocks);
-
-      lut_buffers_bivariate.push_back(shift_blocks_lut_bivariate);
-    }
-  }
-
-  void release(CudaStreams streams) {
-    cuda_synchronize_stream(streams.stream(0), streams.gpu_index(0));
-    local_streams_1.release();
-    local_streams_2.release();
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_rotated, gpu_memory_allocated);
-    for (auto &buffer : lut_buffers_bivariate) {
-      buffer->release(streams);
-      delete buffer;
-    }
-    for (auto &buffer : lut_buffers_univariate) {
-      buffer->release(streams);
-      delete buffer;
-    }
-    lut_buffers_bivariate.clear();
-    lut_buffers_univariate.clear();
-
-    delete tmp_rotated;
-  }
-};
-
-template <typename Torus> struct int_cmux_buffer {
-  int_radix_lut<Torus> *predicate_lut;
-  int_radix_lut<Torus> *message_extract_lut;
-
-  CudaRadixCiphertextFFI *buffer_in;
-  CudaRadixCiphertextFFI *buffer_out;
-  CudaRadixCiphertextFFI *condition_array;
-
-  int_radix_params params;
-  bool allocate_gpu_memory;
-  bool gpu_memory_allocated;
-  int_cmux_buffer(CudaStreams streams,
-                  std::function<Torus(Torus)> predicate_lut_f,
-                  int_radix_params params, uint32_t num_radix_blocks,
-                  bool allocate_gpu_memory, uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-
-    buffer_in = new CudaRadixCiphertextFFI;
-    buffer_out = new CudaRadixCiphertextFFI;
-    condition_array = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), buffer_in,
-        2 * num_radix_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), buffer_out,
-        2 * num_radix_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), condition_array,
-        2 * num_radix_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    auto lut_f = [predicate_lut_f](Torus block, Torus condition) -> Torus {
-      return predicate_lut_f(condition) ? 0 : block;
-    };
-    auto inverted_lut_f = [predicate_lut_f](Torus block,
-                                            Torus condition) -> Torus {
-      return predicate_lut_f(condition) ? block : 0;
-    };
-    auto message_extract_lut_f = [params](Torus x) -> Torus {
-      return x % params.message_modulus;
-    };
-
-    predicate_lut =
-        new int_radix_lut<Torus>(streams, params, 2, 2 * num_radix_blocks,
-                                 allocate_gpu_memory, size_tracker);
-
-    message_extract_lut =
-        new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                 allocate_gpu_memory, size_tracker);
-
-    generate_device_accumulator_bivariate<Torus>(
-        streams.stream(0), streams.gpu_index(0), predicate_lut->get_lut(0, 0),
-        predicate_lut->get_degree(0), predicate_lut->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, inverted_lut_f, gpu_memory_allocated);
-
-    generate_device_accumulator_bivariate<Torus>(
-        streams.stream(0), streams.gpu_index(0), predicate_lut->get_lut(0, 1),
-        predicate_lut->get_degree(1), predicate_lut->get_max_degree(1),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, lut_f, gpu_memory_allocated);
-
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0),
-        message_extract_lut->get_lut(0, 0), message_extract_lut->get_degree(0),
-        message_extract_lut->get_max_degree(0), params.glwe_dimension,
-        params.polynomial_size, params.message_modulus, params.carry_modulus,
-        message_extract_lut_f, gpu_memory_allocated);
-    Torus *h_lut_indexes = predicate_lut->h_lut_indexes;
-    for (int index = 0; index < 2 * num_radix_blocks; index++) {
-      if (index < num_radix_blocks) {
-        h_lut_indexes[index] = 0;
-      } else {
-        h_lut_indexes[index] = 1;
-      }
-    }
-    cuda_memcpy_with_size_tracking_async_to_gpu(
-        predicate_lut->get_lut_indexes(0, 0), h_lut_indexes,
-        2 * num_radix_blocks * sizeof(Torus), streams.stream(0),
-        streams.gpu_index(0), allocate_gpu_memory);
-    auto active_streams_pred = streams.active_gpu_subset(2 * num_radix_blocks);
-    predicate_lut->broadcast_lut(active_streams_pred);
-    auto active_streams_msg = streams.active_gpu_subset(num_radix_blocks);
-    message_extract_lut->broadcast_lut(active_streams_msg);
-  }
-
-  void release(CudaStreams streams) {
-    predicate_lut->release(streams);
-    delete predicate_lut;
-    message_extract_lut->release(streams);
-    delete message_extract_lut;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   buffer_in, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   buffer_out, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   condition_array, gpu_memory_allocated);
-    cuda_synchronize_stream(streams.stream(0), streams.gpu_index(0));
-    delete buffer_in;
-    delete buffer_out;
-    delete condition_array;
-  }
-};
-
-template <typename Torus> struct int_are_all_block_true_buffer {
-  COMPARISON_TYPE op;
-  int_radix_params params;
-
-  CudaRadixCiphertextFFI *tmp_out;
-  CudaRadixCiphertextFFI *tmp_block_accumulated;
-
-  // This map store LUTs that checks the equality between some input and values
-  // of interest in are_all_block_true(), as with max_value (the maximum message
-  // value).
-  int_radix_lut<Torus> *is_max_value;
-  Torus *preallocated_h_lut;
-  bool gpu_memory_allocated;
-
-  int_are_all_block_true_buffer(CudaStreams streams, COMPARISON_TYPE op,
-                                int_radix_params params,
-                                uint32_t num_radix_blocks,
-                                bool allocate_gpu_memory,
-                                uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-    this->params = params;
-    this->op = op;
-
-    Torus total_modulus = params.message_modulus * params.carry_modulus;
-    uint32_t max_value = (total_modulus - 1) / (params.message_modulus - 1);
-
-    int max_chunks = (num_radix_blocks + max_value - 1) / max_value;
-    tmp_out = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_out, num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    tmp_block_accumulated = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_block_accumulated,
-        max_chunks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    is_max_value = new int_radix_lut<Torus>(streams, params, 2, max_chunks,
-                                            allocate_gpu_memory, size_tracker);
-    auto is_max_value_f = [max_value](Torus x) -> Torus {
-      return x == max_value;
-    };
-    preallocated_h_lut = (Torus *)malloc(
-        (params.glwe_dimension + 1) * params.polynomial_size * sizeof(Torus));
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0), is_max_value->get_lut(0, 0),
-        is_max_value->get_degree(0), is_max_value->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, is_max_value_f, gpu_memory_allocated);
-
-    auto active_streams = streams.active_gpu_subset(max_chunks);
-    is_max_value->broadcast_lut(active_streams);
-  }
-
-  void release(CudaStreams streams) {
-    cuda_synchronize_stream(streams.stream(0), streams.gpu_index(0));
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_out, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_block_accumulated, gpu_memory_allocated);
-    is_max_value->release(streams);
-    free(preallocated_h_lut);
-    delete is_max_value;
-    delete tmp_out;
-    delete tmp_block_accumulated;
-  }
-};
-
-template <typename Torus> struct int_comparison_eq_buffer {
-  int_radix_params params;
-  COMPARISON_TYPE op;
-
-  int_radix_lut<Torus> *operator_lut;
-  int_radix_lut<Torus> *is_non_zero_lut;
-  int_radix_lut<Torus> *scalar_comparison_luts;
-
-  int_are_all_block_true_buffer<Torus> *are_all_block_true_buffer;
-  bool gpu_memory_allocated;
-
-  int_comparison_eq_buffer(CudaStreams streams, COMPARISON_TYPE op,
-                           int_radix_params params, uint32_t num_radix_blocks,
-                           bool allocate_gpu_memory, uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-    this->params = params;
-    this->op = op;
-    Torus total_modulus = params.message_modulus * params.carry_modulus;
-
-    are_all_block_true_buffer = new int_are_all_block_true_buffer<Torus>(
-        streams, op, params, num_radix_blocks, allocate_gpu_memory,
-        size_tracker);
-
-    // f(x) -> x == 0
-    auto is_non_zero_lut_f = [total_modulus](Torus x) -> Torus {
-      return (x % total_modulus) != 0;
-    };
-
-    is_non_zero_lut =
-        new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                 allocate_gpu_memory, size_tracker);
-
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0), is_non_zero_lut->get_lut(0, 0),
-        is_non_zero_lut->get_degree(0), is_non_zero_lut->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, is_non_zero_lut_f, gpu_memory_allocated);
-
-    auto active_streams = streams.active_gpu_subset(num_radix_blocks);
-    is_non_zero_lut->broadcast_lut(active_streams);
-
-    // Scalar may have up to num_radix_blocks blocks
-    scalar_comparison_luts = new int_radix_lut<Torus>(
-        streams, params, total_modulus, num_radix_blocks, allocate_gpu_memory,
-        size_tracker);
-
-    // Operator LUT
-    auto operator_f = [op](Torus lhs, Torus rhs) -> Torus {
-      if (op == COMPARISON_TYPE::EQ) {
-        return (lhs == rhs);
-      } else if (op == COMPARISON_TYPE::NE) {
-        return (lhs != rhs);
-      } else {
-        // For signed scalar comparisons we check equality with zero
-        return (lhs == rhs);
-      }
-    };
-    for (int i = 0; i < total_modulus; i++) {
-      auto lut_f = [i, operator_f](Torus x) -> Torus {
-        return operator_f(i, x);
-      };
-
-      generate_device_accumulator<Torus>(
-          streams.stream(0), streams.gpu_index(0),
-          scalar_comparison_luts->get_lut(0, i),
-          scalar_comparison_luts->get_degree(i),
-          scalar_comparison_luts->get_max_degree(i), params.glwe_dimension,
-          params.polynomial_size, params.message_modulus, params.carry_modulus,
-          lut_f, gpu_memory_allocated);
-    }
-    scalar_comparison_luts->broadcast_lut(active_streams);
-    if (op == COMPARISON_TYPE::EQ || op == COMPARISON_TYPE::NE) {
-      operator_lut =
-          new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                   allocate_gpu_memory, size_tracker);
-
-      generate_device_accumulator_bivariate<Torus>(
-          streams.stream(0), streams.gpu_index(0), operator_lut->get_lut(0, 0),
-          operator_lut->get_degree(0), operator_lut->get_max_degree(0),
-          params.glwe_dimension, params.polynomial_size, params.message_modulus,
-          params.carry_modulus, operator_f, gpu_memory_allocated);
-
-      operator_lut->broadcast_lut(active_streams);
-    } else {
-      operator_lut = nullptr;
-    }
-  }
-
-  void release(CudaStreams streams) {
-    if (op == COMPARISON_TYPE::EQ || op == COMPARISON_TYPE::NE) {
-      PANIC_IF_FALSE(operator_lut != nullptr,
-                     "Cuda error: no operator lut was created");
-      operator_lut->release(streams);
-      delete operator_lut;
-      operator_lut = nullptr;
-    }
-    is_non_zero_lut->release(streams);
-    delete is_non_zero_lut;
-    is_non_zero_lut = nullptr;
-    are_all_block_true_buffer->release(streams);
-    delete are_all_block_true_buffer;
-    are_all_block_true_buffer = nullptr;
-    scalar_comparison_luts->release(streams);
-    delete scalar_comparison_luts;
-    scalar_comparison_luts = nullptr;
-  }
-};
-
-template <typename Torus> struct int_tree_sign_reduction_buffer {
-  int_radix_params params;
-
-  std::function<Torus(Torus, Torus)> block_selector_f;
-
-  int_radix_lut<Torus> *tree_inner_leaf_lut;
-  int_radix_lut<Torus> *tree_last_leaf_lut;
-
-  int_radix_lut<Torus> *tree_last_leaf_scalar_lut;
-
-  Torus *preallocated_h_lut;
-  CudaRadixCiphertextFFI *tmp_x;
-  CudaRadixCiphertextFFI *tmp_y;
-  bool gpu_memory_allocated;
-
-  int_tree_sign_reduction_buffer(CudaStreams streams,
-                                 std::function<Torus(Torus)> operator_f,
-                                 int_radix_params params,
-                                 uint32_t num_radix_blocks,
-                                 bool allocate_gpu_memory,
-                                 uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-    this->params = params;
-
-    block_selector_f = [](Torus msb, Torus lsb) -> Torus {
-      if (msb == IS_EQUAL) // EQUAL
-        return lsb;
-      else
-        return msb;
-    };
-
-    tmp_x = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_x, num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    tmp_y = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_y, num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    // LUTs
-    tree_inner_leaf_lut =
-        new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                 allocate_gpu_memory, size_tracker);
-
-    tree_last_leaf_lut = new int_radix_lut<Torus>(
-        streams, params, 1, 1, allocate_gpu_memory, size_tracker);
-
-    preallocated_h_lut = (Torus *)malloc(
-        (params.glwe_dimension + 1) * params.polynomial_size * sizeof(Torus));
-
-    tree_last_leaf_scalar_lut = new int_radix_lut<Torus>(
-        streams, params, 1, 1, allocate_gpu_memory, size_tracker);
-
-    generate_device_accumulator_bivariate<Torus>(
-        streams.stream(0), streams.gpu_index(0),
-        tree_inner_leaf_lut->get_lut(0, 0), tree_inner_leaf_lut->get_degree(0),
-        tree_inner_leaf_lut->get_max_degree(0), params.glwe_dimension,
-        params.polynomial_size, params.message_modulus, params.carry_modulus,
-        block_selector_f, gpu_memory_allocated);
-    auto active_streams = streams.active_gpu_subset(num_radix_blocks);
-    tree_inner_leaf_lut->broadcast_lut(active_streams);
-  }
-
-  void release(CudaStreams streams) {
-    cuda_synchronize_stream(streams.stream(0), streams.gpu_index(0));
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_x, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_y, gpu_memory_allocated);
-    tree_inner_leaf_lut->release(streams);
-    delete tree_inner_leaf_lut;
-    tree_last_leaf_lut->release(streams);
-    delete tree_last_leaf_lut;
-    tree_last_leaf_scalar_lut->release(streams);
-    delete tree_last_leaf_scalar_lut;
-
-    delete tmp_x;
-    delete tmp_y;
-    free(preallocated_h_lut);
-  }
-};
-
-template <typename Torus> struct int_comparison_diff_buffer {
-  int_radix_params params;
-  COMPARISON_TYPE op;
-
-  CudaRadixCiphertextFFI *tmp_packed;
-
-  std::function<Torus(Torus)> operator_f;
-
-  int_tree_sign_reduction_buffer<Torus> *tree_buffer;
-
-  CudaRadixCiphertextFFI *tmp_signs_a;
-  CudaRadixCiphertextFFI *tmp_signs_b;
-  int_radix_lut<Torus> *reduce_signs_lut;
-  bool gpu_memory_allocated;
-  Torus *preallocated_h_lut1;
-  Torus *preallocated_h_lut2;
-  int_comparison_diff_buffer(CudaStreams streams, COMPARISON_TYPE op,
-                             int_radix_params params, uint32_t num_radix_blocks,
-                             bool allocate_gpu_memory, uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-    this->params = params;
-    this->op = op;
-
-    operator_f = [op](Torus x) -> Torus {
-      switch (op) {
-      case GT:
-        return x == IS_SUPERIOR;
-      case GE:
-        return (x == IS_SUPERIOR) || (x == IS_EQUAL);
-      case LT:
-        return x == IS_INFERIOR;
-      case LE:
-        return (x == IS_INFERIOR) || (x == IS_EQUAL);
-      default:
-        PANIC("Cuda error (comparisons): unknown comparison type")
-      }
-    };
-
-    tmp_packed = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_packed, num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-    tree_buffer = new int_tree_sign_reduction_buffer<Torus>(
-        streams, operator_f, params, num_radix_blocks, allocate_gpu_memory,
-        size_tracker);
-    tmp_signs_a = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_signs_a, num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    tmp_signs_b = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_signs_b, num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    // LUTs
-    reduce_signs_lut =
-        new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                 allocate_gpu_memory, size_tracker);
-    preallocated_h_lut1 = (Torus *)malloc(
-        (params.glwe_dimension + 1) * params.polynomial_size * sizeof(Torus));
-    preallocated_h_lut2 = (Torus *)malloc(
-        (params.glwe_dimension + 1) * params.polynomial_size * sizeof(Torus));
-  }
-
-  void release(CudaStreams streams) {
-    cuda_synchronize_stream(streams.stream(0), streams.gpu_index(0));
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_packed, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_signs_a, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_signs_b, gpu_memory_allocated);
-    tree_buffer->release(streams);
-    delete tree_buffer;
-    reduce_signs_lut->release(streams);
-    delete reduce_signs_lut;
-
-    delete tmp_packed;
-    delete tmp_signs_a;
-    delete tmp_signs_b;
-    free(preallocated_h_lut1);
-    free(preallocated_h_lut2);
-  }
-};
-
-template <typename Torus> struct int_comparison_buffer {
-  COMPARISON_TYPE op;
-
-  int_radix_params params;
-
-  //////////////////
-  int_radix_lut<Torus> *identity_lut;
-  std::function<Torus(Torus)> identity_lut_f;
-
-  int_radix_lut<Torus> *is_zero_lut;
-
-  int_comparison_eq_buffer<Torus> *eq_buffer;
-  int_comparison_diff_buffer<Torus> *diff_buffer;
-
-  CudaRadixCiphertextFFI *tmp_block_comparisons;
-  CudaRadixCiphertextFFI *tmp_lwe_array_out;
-  CudaRadixCiphertextFFI *tmp_trivial_sign_block;
-
-  // Scalar EQ / NE
-  CudaRadixCiphertextFFI *tmp_packed_input;
-
-  // Max Min
-  int_cmux_buffer<Torus> *cmux_buffer;
-
-  // Signed LUT
-  int_radix_lut<Torus> *signed_lut;
-  bool is_signed;
-
-  // Used for scalar comparisons
-  int_radix_lut<Torus> *signed_msb_lut;
-  CudaStreams lsb_streams;
-  CudaStreams msb_streams;
-  bool gpu_memory_allocated;
-  Torus *preallocated_h_lut;
-
-  int_comparison_buffer(CudaStreams streams, COMPARISON_TYPE op,
-                        int_radix_params params, uint32_t num_radix_blocks,
-                        bool is_signed, bool allocate_gpu_memory,
-                        uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-    this->params = params;
-    this->op = op;
-    this->is_signed = is_signed;
-
-    auto active_streams = streams.active_gpu_subset(num_radix_blocks);
-
-    identity_lut_f = [](Torus x) -> Torus { return x; };
-
-    lsb_streams.create_on_same_gpus(active_streams);
-    msb_streams.create_on_same_gpus(active_streams);
-
-    // +1 to have space for signed comparison
-    tmp_lwe_array_out = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_lwe_array_out,
-        num_radix_blocks + 1, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    tmp_packed_input = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_packed_input,
-        2 * num_radix_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    // Block comparisons
-    tmp_block_comparisons = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_block_comparisons,
-        num_radix_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    // Cleaning LUT
-    identity_lut =
-        new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                 allocate_gpu_memory, size_tracker);
-
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0), identity_lut->get_lut(0, 0),
-        identity_lut->get_degree(0), identity_lut->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, identity_lut_f, gpu_memory_allocated);
-    identity_lut->broadcast_lut(active_streams);
-
-    uint32_t total_modulus = params.message_modulus * params.carry_modulus;
-    auto is_zero_f = [total_modulus](Torus x) -> Torus {
-      return (x % total_modulus) == 0;
-    };
-
-    is_zero_lut = new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                           allocate_gpu_memory, size_tracker);
-
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0), is_zero_lut->get_lut(0, 0),
-        is_zero_lut->get_degree(0), is_zero_lut->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, is_zero_f, gpu_memory_allocated);
-
-    is_zero_lut->broadcast_lut(active_streams);
-
-    switch (op) {
-    case COMPARISON_TYPE::MAX:
-    case COMPARISON_TYPE::MIN:
-      cmux_buffer = new int_cmux_buffer<Torus>(
-          streams,
-          [op](Torus x) -> Torus {
-            if (op == COMPARISON_TYPE::MAX)
-              return (x == IS_SUPERIOR);
-            else
-              return (x == IS_INFERIOR);
-          },
-          params, num_radix_blocks, allocate_gpu_memory, size_tracker);
-    case COMPARISON_TYPE::GT:
-    case COMPARISON_TYPE::GE:
-    case COMPARISON_TYPE::LT:
-    case COMPARISON_TYPE::LE:
-      diff_buffer = new int_comparison_diff_buffer<Torus>(
-          streams, op, params, num_radix_blocks, allocate_gpu_memory,
-          size_tracker);
-    case COMPARISON_TYPE::EQ:
-    case COMPARISON_TYPE::NE:
-      eq_buffer = new int_comparison_eq_buffer<Torus>(
-          streams, op, params, num_radix_blocks, allocate_gpu_memory,
-          size_tracker);
-      break;
-    default:
-      PANIC("Unsupported comparison operation.")
-    }
-
-    if (is_signed) {
-
-      tmp_trivial_sign_block = new CudaRadixCiphertextFFI;
-      create_zero_radix_ciphertext_async<Torus>(
-          streams.stream(0), streams.gpu_index(0), tmp_trivial_sign_block, 1,
-          params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-      signed_lut = new int_radix_lut<Torus>(streams, params, 1, 1,
-                                            allocate_gpu_memory, size_tracker);
-      signed_msb_lut = new int_radix_lut<Torus>(
-          streams, params, 1, 1, allocate_gpu_memory, size_tracker);
-
-      auto message_modulus = (int)params.message_modulus;
-      uint32_t sign_bit_pos = log2(message_modulus) - 1;
-      std::function<Torus(Torus, Torus)> signed_lut_f =
-          [sign_bit_pos](Torus x, Torus y) -> Torus {
-        auto x_sign_bit = x >> sign_bit_pos;
-        auto y_sign_bit = y >> sign_bit_pos;
-
-        // The block that has its sign bit set is going
-        // to be ordered as 'greater' by the cmp fn.
-        // However, we are dealing with signed number,
-        // so in reality, it is the smaller of the two.
-        // i.e the cmp result is reversed
-        if (x_sign_bit == y_sign_bit) {
-          // Both have either sign bit set or unset,
-          // cmp will give correct result
-          if (x < y)
-            return (Torus)(IS_INFERIOR);
-          else if (x == y)
-            return (Torus)(IS_EQUAL);
-          else
-            return (Torus)(IS_SUPERIOR);
-        } else {
-          if (x < y)
-            return (Torus)(IS_SUPERIOR);
-          else if (x == y)
-            return (Torus)(IS_EQUAL);
-          else
-            return (Torus)(IS_INFERIOR);
-        }
-        PANIC("Cuda error: sign_lut creation failed due to wrong function.")
-      };
-
-      generate_device_accumulator_bivariate<Torus>(
-          streams.stream(0), streams.gpu_index(0), signed_lut->get_lut(0, 0),
-          signed_lut->get_degree(0), signed_lut->get_max_degree(0),
-          params.glwe_dimension, params.polynomial_size, params.message_modulus,
-          params.carry_modulus, signed_lut_f, gpu_memory_allocated);
-      auto active_streams = streams.active_gpu_subset(1);
-      signed_lut->broadcast_lut(active_streams);
-    }
-    preallocated_h_lut = (Torus *)malloc(
-        (params.glwe_dimension + 1) * params.polynomial_size * sizeof(Torus));
-  }
-
-  void release(CudaStreams streams) {
-    switch (op) {
-    case COMPARISON_TYPE::MAX:
-    case COMPARISON_TYPE::MIN:
-      cmux_buffer->release(streams);
-      delete (cmux_buffer);
-    case COMPARISON_TYPE::GT:
-    case COMPARISON_TYPE::GE:
-    case COMPARISON_TYPE::LT:
-    case COMPARISON_TYPE::LE:
-      diff_buffer->release(streams);
-      delete (diff_buffer);
-    case COMPARISON_TYPE::EQ:
-    case COMPARISON_TYPE::NE:
-      eq_buffer->release(streams);
-      delete (eq_buffer);
-      break;
-    default:
-      PANIC("Unsupported comparison operation.")
-    }
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_lwe_array_out, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_block_comparisons, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_packed_input, gpu_memory_allocated);
-    identity_lut->release(streams);
-    delete identity_lut;
-    is_zero_lut->release(streams);
-    delete is_zero_lut;
-    delete tmp_lwe_array_out;
-    delete tmp_block_comparisons;
-    delete tmp_packed_input;
-
-    if (is_signed) {
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     tmp_trivial_sign_block,
-                                     gpu_memory_allocated);
-      signed_lut->release(streams);
-      delete signed_lut;
-      signed_msb_lut->release(streams);
-      delete signed_msb_lut;
-      delete tmp_trivial_sign_block;
-    }
-    cuda_synchronize_stream(streams.stream(0), streams.gpu_index(0));
-    lsb_streams.release();
-    msb_streams.release();
-    free(preallocated_h_lut);
-  }
-};
-
-// used only when 4 gpus are available
-template <typename Torus> struct unsigned_int_div_rem_2_2_memory {
-  int_radix_params params;
-  bool gpu_memory_allocated;
-
-  // memory objects for other operations
-  int_borrow_prop_memory<Torus> *overflow_sub_mem_1;
-  int_borrow_prop_memory<Torus> *overflow_sub_mem_2;
-  int_borrow_prop_memory<Torus> *overflow_sub_mem_3;
-  int_comparison_buffer<Torus> *comparison_buffer_1;
-  int_comparison_buffer<Torus> *comparison_buffer_2;
-  int_comparison_buffer<Torus> *comparison_buffer_3;
-  int_sub_and_propagate<Torus> *sub_and_propagate_mem;
-  int_bitop_buffer<Torus> *bitor_mem_1;
-  int_bitop_buffer<Torus> *bitor_mem_2;
-  int_bitop_buffer<Torus> *bitor_mem_3;
-  int_logical_scalar_shift_buffer<Torus> *shift_mem;
-
-  // lookup tables
-  int_radix_lut<Torus> *message_extract_lut_1;
-  int_radix_lut<Torus> *message_extract_lut_2;
-  int_radix_lut<Torus> *zero_out_if_not_1_lut_1;
-  int_radix_lut<Torus> *zero_out_if_not_1_lut_2;
-  int_radix_lut<Torus> *zero_out_if_not_2_lut_1;
-  int_radix_lut<Torus> *zero_out_if_not_2_lut_2;
-  int_radix_lut<Torus> *quotient_lut_1;
-  int_radix_lut<Torus> *quotient_lut_2;
-  int_radix_lut<Torus> *quotient_lut_3;
-
-  // sub streams
-  CudaStreams sub_streams_1;
-
-  // temporary device buffers
-  CudaRadixCiphertextFFI *d1;                  // num_blocks + 1
-  CudaRadixCiphertextFFI *d2;                  // num_blocks + 1
-  CudaRadixCiphertextFFI *d3;                  // num_blocks + 1
-  CudaRadixCiphertextFFI *tmp_gpu_0;           // num_blocks + 1
-  CudaRadixCiphertextFFI *tmp_gpu_1;           // num_blocks + 1
-  CudaRadixCiphertextFFI *tmp_gpu_2;           // num_blocks + 1
-  CudaRadixCiphertextFFI *tmp_gpu_3;           // num_blocks + 1
-  CudaRadixCiphertextFFI *divisor_gpu_1;       // num_blocks
-  CudaRadixCiphertextFFI *divisor_gpu_2;       // num_blocks
-  CudaRadixCiphertextFFI *remainder_gpu_1;     // num_blocks
-  CudaRadixCiphertextFFI *remainder_gpu_2;     // num_blocks
-  CudaRadixCiphertextFFI *remainder_gpu_3;     // num_blocks
-  CudaRadixCiphertextFFI *low1;                // num_blocks
-  CudaRadixCiphertextFFI *low2;                // num_blocks
-  CudaRadixCiphertextFFI *low3;                // num_blocks
-  CudaRadixCiphertextFFI *rem0;                // num_blocks
-  CudaRadixCiphertextFFI *rem1;                // num_blocks
-  CudaRadixCiphertextFFI *rem2;                // num_blocks
-  CudaRadixCiphertextFFI *rem3;                // num_blocks
-  CudaRadixCiphertextFFI *sub_result_1;        // num_blocks
-  CudaRadixCiphertextFFI *sub_result_2;        // num_blocks
-  CudaRadixCiphertextFFI *sub_result_3;        // num_blocks
-  CudaRadixCiphertextFFI *sub_1_overflowed;    // num_blocks
-  CudaRadixCiphertextFFI *sub_2_overflowed;    // num_blocks
-  CudaRadixCiphertextFFI *sub_3_overflowed;    // num_blocks
-  CudaRadixCiphertextFFI *comparison_blocks_1; // num_blocks
-  CudaRadixCiphertextFFI *comparison_blocks_2; // num_blocks
-  CudaRadixCiphertextFFI *comparison_blocks_3; // num_blocks
-  CudaRadixCiphertextFFI *cmp_1;               // boolean block
-  CudaRadixCiphertextFFI *cmp_2;               // boolean block
-  CudaRadixCiphertextFFI *cmp_3;               // boolean block
-  CudaRadixCiphertextFFI *c0;                  // single block
-  CudaRadixCiphertextFFI *q1;                  // single block
-  CudaRadixCiphertextFFI *q2;                  // single block
-  CudaRadixCiphertextFFI *q3;                  // single block
-
-  Torus *h_buffer; // used for memory copies
-
-  Torus **first_indexes_for_overflow_sub_gpu_0;
-  Torus **second_indexes_for_overflow_sub_gpu_0;
-  Torus **scalars_for_overflow_sub_gpu_0;
-
-  Torus **first_indexes_for_overflow_sub_gpu_1;
-  Torus **second_indexes_for_overflow_sub_gpu_1;
-  Torus **scalars_for_overflow_sub_gpu_1;
-
-  Torus **first_indexes_for_overflow_sub_gpu_2;
-  Torus **second_indexes_for_overflow_sub_gpu_2;
-  Torus **scalars_for_overflow_sub_gpu_2;
-
-  cudaEvent_t create_indexes_done;
-
-  uint32_t max_indexes_to_erase;
-  uint64_t tmp_size_tracker = 0;
-
-  // allocate and initialize if needed, temporary arrays used to calculate
-  // cuda integer div_rem_2_2 operation
-  void init_temporary_buffers(CudaStreams streams, uint32_t num_blocks,
-                              bool allocate_gpu_memory,
-                              uint64_t &size_tracker) {
-
-    // more than one block temporary arrays
-    tmp_gpu_0 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_gpu_0, num_blocks + 1,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    d3 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), d3, num_blocks + 1,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    low3 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), low3, num_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    rem3 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), rem3, num_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    sub_result_1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), sub_result_1, num_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    sub_1_overflowed = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), sub_1_overflowed, 1,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    comparison_blocks_1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), comparison_blocks_1,
-        num_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-    cmp_1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), cmp_1, 1,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    q3 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), q3, 1,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-    tmp_gpu_1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(1), streams.gpu_index(1), tmp_gpu_1, num_blocks + 1,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    d2 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(1), streams.gpu_index(1), d2, num_blocks + 1,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    low2 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(1), streams.gpu_index(1), low2, num_blocks,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    rem2 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(1), streams.gpu_index(1), rem2, num_blocks,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    divisor_gpu_1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(1), streams.gpu_index(1), divisor_gpu_1, num_blocks,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    remainder_gpu_1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(1), streams.gpu_index(1), remainder_gpu_1, num_blocks,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    sub_result_2 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(1), streams.gpu_index(1), sub_result_2, num_blocks,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    sub_2_overflowed = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(1), streams.gpu_index(1), sub_2_overflowed, 1,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    comparison_blocks_2 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(1), streams.gpu_index(1), comparison_blocks_2,
-        num_blocks, params.big_lwe_dimension, tmp_size_tracker,
-        allocate_gpu_memory);
-    cmp_2 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(1), streams.gpu_index(1), cmp_2, 1,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    q2 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(1), streams.gpu_index(1), q2, 1,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-
-    tmp_gpu_2 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(2), streams.gpu_index(2), tmp_gpu_2, num_blocks + 1,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    d1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(2), streams.gpu_index(2), d1, num_blocks + 1,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    low1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(2), streams.gpu_index(2), low1, num_blocks,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    rem1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(2), streams.gpu_index(2), rem1, num_blocks,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    divisor_gpu_2 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(2), streams.gpu_index(2), divisor_gpu_2, num_blocks,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    remainder_gpu_2 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(2), streams.gpu_index(2), remainder_gpu_2, num_blocks,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    sub_result_3 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(2), streams.gpu_index(2), sub_result_3, num_blocks,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    sub_3_overflowed = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(2), streams.gpu_index(2), sub_3_overflowed, 1,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    comparison_blocks_3 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(2), streams.gpu_index(2), comparison_blocks_3,
-        num_blocks, params.big_lwe_dimension, tmp_size_tracker,
-        allocate_gpu_memory);
-    cmp_3 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(2), streams.gpu_index(2), cmp_3, 1,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    q1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(2), streams.gpu_index(2), q1, 1,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-
-    tmp_gpu_3 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(3), streams.gpu_index(3), tmp_gpu_3, num_blocks + 1,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    rem0 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(3), streams.gpu_index(3), rem0, num_blocks,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    remainder_gpu_3 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(3), streams.gpu_index(3), remainder_gpu_3, num_blocks,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-    c0 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(3), streams.gpu_index(3), c0, 1,
-        params.big_lwe_dimension, tmp_size_tracker, allocate_gpu_memory);
-  }
-
-  // initialize lookup tables for div_rem_2_2 operation
-  void init_lookup_tables(CudaStreams streams, uint32_t num_blocks,
-                          bool allocate_gpu_memory, uint64_t &size_tracker) {
-
-    zero_out_if_not_1_lut_1 =
-        new int_radix_lut<Torus>(streams.get_ith(0), params, 1, num_blocks,
-                                 allocate_gpu_memory, size_tracker);
-
-    zero_out_if_not_2_lut_1 =
-        new int_radix_lut<Torus>(streams.get_ith(1), params, 1, num_blocks,
-                                 allocate_gpu_memory, tmp_size_tracker);
-
-    zero_out_if_not_2_lut_2 =
-        new int_radix_lut<Torus>(streams.get_ith(2), params, 1, num_blocks,
-                                 allocate_gpu_memory, tmp_size_tracker);
-
-    zero_out_if_not_1_lut_2 =
-        new int_radix_lut<Torus>(streams.get_ith(3), params, 1, num_blocks,
-                                 allocate_gpu_memory, tmp_size_tracker);
-
-    auto zero_out_if_not_1_lut_f = [](Torus x) -> Torus {
-      Torus block = x / 2;
-      bool condition = (x & 1) == 1;
-      return block * (Torus)condition;
-    };
-    auto zero_out_if_not_2_lut_f = [](Torus x) -> Torus {
-      Torus block = x / 3;
-      bool condition = (x % 3) == 2;
-      return block * (Torus)condition;
-    };
-
-    int_radix_lut<Torus> *luts[2] = {zero_out_if_not_1_lut_1,
-                                     zero_out_if_not_1_lut_2};
-    size_t lut_gpu_indexes[2] = {0, 3};
-    for (int j = 0; j < 2; j++) {
-      generate_device_accumulator<Torus>(
-          streams.stream(lut_gpu_indexes[j]),
-          streams.gpu_index(lut_gpu_indexes[j]), luts[j]->get_lut(0, 0),
-          luts[j]->get_degree(0), luts[j]->get_max_degree(0),
-          params.glwe_dimension, params.polynomial_size, params.message_modulus,
-          params.carry_modulus, zero_out_if_not_1_lut_f, gpu_memory_allocated);
-    }
-
-    luts[0] = zero_out_if_not_2_lut_1;
-    luts[1] = zero_out_if_not_2_lut_2;
-    lut_gpu_indexes[0] = 1;
-    lut_gpu_indexes[1] = 2;
-    for (int j = 0; j < 2; j++) {
-      generate_device_accumulator<Torus>(
-          streams.stream(lut_gpu_indexes[j]),
-          streams.gpu_index(lut_gpu_indexes[j]), luts[j]->get_lut(0, 0),
-          luts[j]->get_degree(0), luts[j]->get_max_degree(0),
-          params.glwe_dimension, params.polynomial_size, params.message_modulus,
-          params.carry_modulus, zero_out_if_not_2_lut_f, gpu_memory_allocated);
-    }
-
-    quotient_lut_1 =
-        new int_radix_lut<Torus>(streams.get_ith(2), params, 1, 1,
-                                 allocate_gpu_memory, tmp_size_tracker);
-    quotient_lut_2 =
-        new int_radix_lut<Torus>(streams.get_ith(1), params, 1, 1,
-                                 allocate_gpu_memory, tmp_size_tracker);
-    quotient_lut_3 = new int_radix_lut<Torus>(
-        streams.get_ith(0), params, 1, 1, allocate_gpu_memory, size_tracker);
-
-    auto quotient_lut_1_f = [](Torus cond) -> Torus {
-      return (Torus)(cond == 2);
-    };
-    auto quotient_lut_2_f = [](Torus cond) -> Torus {
-      return (Torus)((cond == 2) * 2);
-    };
-    auto quotient_lut_3_f = [](Torus cond) -> Torus { return cond * 3; };
-
-    generate_device_accumulator<Torus>(
-        streams.stream(2), streams.gpu_index(2), quotient_lut_1->get_lut(0, 0),
-        quotient_lut_1->get_degree(0), quotient_lut_1->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, quotient_lut_1_f, gpu_memory_allocated);
-    generate_device_accumulator<Torus>(
-        streams.stream(1), streams.gpu_index(1), quotient_lut_2->get_lut(0, 0),
-        quotient_lut_2->get_degree(0), quotient_lut_2->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, quotient_lut_2_f, gpu_memory_allocated);
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0), quotient_lut_3->get_lut(0, 0),
-        quotient_lut_3->get_degree(0), quotient_lut_3->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, quotient_lut_3_f, gpu_memory_allocated);
-
-    message_extract_lut_1 = new int_radix_lut<Torus>(
-        streams, params, 1, num_blocks, allocate_gpu_memory, size_tracker);
-    message_extract_lut_2 = new int_radix_lut<Torus>(
-        streams, params, 1, num_blocks, allocate_gpu_memory, size_tracker);
-
-    auto message_modulus = params.message_modulus;
-    auto lut_f_message_extract = [message_modulus](Torus x) -> Torus {
-      return x % message_modulus;
-    };
-
-    luts[0] = message_extract_lut_1;
-    luts[1] = message_extract_lut_2;
-
-    for (int j = 0; j < 2; j++) {
-      generate_device_accumulator<Torus>(
-          streams.stream(0), streams.gpu_index(0), luts[j]->get_lut(0, 0),
-          luts[j]->get_degree(0), luts[j]->get_max_degree(0),
-          params.glwe_dimension, params.polynomial_size, params.message_modulus,
-          params.carry_modulus, lut_f_message_extract, gpu_memory_allocated);
-      auto active_streams = streams.active_gpu_subset(num_blocks);
-      luts[j]->broadcast_lut(active_streams);
-    }
-  }
-
-  unsigned_int_div_rem_2_2_memory(CudaStreams streams, int_radix_params params,
-                                  uint32_t num_blocks, bool allocate_gpu_memory,
-                                  uint64_t &size_tracker) {
-    if (streams.count() < 4) {
-      PANIC("GPU count should be greater than 4m when using div_rem_2_2");
-    }
-
-    if (params.message_modulus != 4 || params.carry_modulus != 4) {
-      PANIC("Only message_modulus == 4 && carry_modulus == 4 parameters are "
-            "supported");
-    }
-
-    this->params = params;
-    gpu_memory_allocated = allocate_gpu_memory;
-
-    sub_and_propagate_mem = new int_sub_and_propagate<Torus>(
-        streams.get_ith(0), params, num_blocks + 1, outputFlag::FLAG_NONE,
-        allocate_gpu_memory, size_tracker);
-
-    shift_mem = new int_logical_scalar_shift_buffer<Torus>(
-        streams.get_ith(1), SHIFT_OR_ROTATE_TYPE::LEFT_SHIFT, params,
-        2 * num_blocks, allocate_gpu_memory, tmp_size_tracker);
-
-    uint32_t compute_overflow = 1;
-    overflow_sub_mem_1 = new int_borrow_prop_memory<Torus>(
-        streams.get_ith(0), params, num_blocks, compute_overflow,
-        allocate_gpu_memory, size_tracker);
-    overflow_sub_mem_2 = new int_borrow_prop_memory<Torus>(
-        streams.get_ith(1), params, num_blocks, compute_overflow,
-        allocate_gpu_memory, tmp_size_tracker);
-    overflow_sub_mem_3 = new int_borrow_prop_memory<Torus>(
-        streams.get_ith(2), params, num_blocks, compute_overflow,
-        allocate_gpu_memory, tmp_size_tracker);
-    uint32_t group_size = overflow_sub_mem_1->group_size;
-    bool use_seq = overflow_sub_mem_1->prop_simu_group_carries_mem
-                       ->use_sequential_algorithm_to_resolve_group_carries;
-
-    cuda_set_device(0);
-    cudaEventCreateWithFlags(&create_indexes_done, cudaEventDisableTiming);
-    create_indexes_for_overflow_sub(streams.get_ith(0), num_blocks, group_size,
-                                    use_seq, allocate_gpu_memory, size_tracker);
-    cudaEventRecord(create_indexes_done, streams.stream(0));
-    cuda_set_device(1);
-    cudaStreamWaitEvent(streams.stream(1), create_indexes_done, 0);
-    cuda_set_device(2);
-    cudaStreamWaitEvent(streams.stream(2), create_indexes_done, 0);
-
-    scatter_indexes_for_overflowing_sub(
-        streams.stream(1), streams.gpu_index(1),
-        &first_indexes_for_overflow_sub_gpu_1,
-        &second_indexes_for_overflow_sub_gpu_1, &scalars_for_overflow_sub_gpu_1,
-        num_blocks, allocate_gpu_memory, tmp_size_tracker);
-    scatter_indexes_for_overflowing_sub(
-        streams.stream(2), streams.gpu_index(2),
-        &first_indexes_for_overflow_sub_gpu_2,
-        &second_indexes_for_overflow_sub_gpu_2, &scalars_for_overflow_sub_gpu_2,
-        num_blocks, allocate_gpu_memory, tmp_size_tracker);
-
-    comparison_buffer_1 = new int_comparison_buffer<Torus>(
-        streams.get_ith(0), COMPARISON_TYPE::EQ, params, num_blocks, false,
-        allocate_gpu_memory, size_tracker);
-    comparison_buffer_2 = new int_comparison_buffer<Torus>(
-        streams.get_ith(1), COMPARISON_TYPE::EQ, params, num_blocks, false,
-        allocate_gpu_memory, tmp_size_tracker);
-    comparison_buffer_3 = new int_comparison_buffer<Torus>(
-        streams.get_ith(2), COMPARISON_TYPE::EQ, params, num_blocks, false,
-        allocate_gpu_memory, tmp_size_tracker);
-    bitor_mem_1 = new int_bitop_buffer<Torus>(
-        streams.get_ith(0), BITOP_TYPE::BITOR, params, num_blocks,
-        allocate_gpu_memory, size_tracker);
-    bitor_mem_2 = new int_bitop_buffer<Torus>(
-        streams.get_ith(1), BITOP_TYPE::BITOR, params, num_blocks,
-        allocate_gpu_memory, tmp_size_tracker);
-    bitor_mem_3 = new int_bitop_buffer<Torus>(
-        streams.get_ith(2), BITOP_TYPE::BITOR, params, num_blocks,
-        allocate_gpu_memory, tmp_size_tracker);
-
-    init_lookup_tables(streams, num_blocks, allocate_gpu_memory, size_tracker);
-    init_temporary_buffers(streams, num_blocks, allocate_gpu_memory,
-                           size_tracker);
-
-    sub_streams_1.create_on_same_gpus(streams);
-  }
-
-  void scatter_indexes_for_overflowing_sub(
-      cudaStream_t const stream, size_t gpu_index, Torus ***first_indexes_ptr,
-      Torus ***second_indexes_ptr, Torus ***scalars_ptr, uint32_t num_blocks,
-      bool allocate_gpu_memory, uint64_t &size_tracker) {
-
-    auto first_indexes = (Torus **)malloc(num_blocks * sizeof(Torus *));
-    auto second_indexes = (Torus **)malloc(num_blocks * sizeof(Torus *));
-    auto scalars = (Torus **)malloc(num_blocks * sizeof(Torus *));
-
-    for (int nb = 1; nb <= num_blocks; nb++) {
-      first_indexes[nb - 1] = (Torus *)cuda_malloc_with_size_tracking_async(
-          nb * sizeof(Torus), stream, gpu_index, size_tracker,
-          allocate_gpu_memory);
-      second_indexes[nb - 1] = (Torus *)cuda_malloc_with_size_tracking_async(
-          nb * sizeof(Torus), stream, gpu_index, size_tracker,
-          allocate_gpu_memory);
-      scalars[nb - 1] = (Torus *)cuda_malloc_with_size_tracking_async(
-          nb * sizeof(Torus), stream, gpu_index, size_tracker,
-          allocate_gpu_memory);
-
-      cuda_memcpy_with_size_tracking_async_gpu_to_gpu(
-          first_indexes[nb - 1], first_indexes_for_overflow_sub_gpu_0[nb - 1],
-          nb * sizeof(Torus), stream, gpu_index, allocate_gpu_memory);
-      cuda_memcpy_with_size_tracking_async_gpu_to_gpu(
-          second_indexes[nb - 1], second_indexes_for_overflow_sub_gpu_0[nb - 1],
-          nb * sizeof(Torus), stream, gpu_index, allocate_gpu_memory);
-      cuda_memcpy_with_size_tracking_async_gpu_to_gpu(
-          scalars[nb - 1], scalars_for_overflow_sub_gpu_0[nb - 1],
-          nb * sizeof(Torus), stream, gpu_index, allocate_gpu_memory);
-      *first_indexes_ptr = first_indexes;
-      *second_indexes_ptr = second_indexes;
-      *scalars_ptr = scalars;
-    }
-  }
-
-  void create_indexes_for_overflow_sub(CudaStreams streams, uint32_t num_blocks,
-                                       uint32_t group_size, bool use_seq,
-                                       bool allocate_gpu_memory,
-                                       uint64_t &size_tracker) {
-    max_indexes_to_erase = num_blocks;
-
-    first_indexes_for_overflow_sub_gpu_0 =
-        (Torus **)malloc(num_blocks * sizeof(Torus *));
-    second_indexes_for_overflow_sub_gpu_0 =
-        (Torus **)malloc(num_blocks * sizeof(Torus *));
-    scalars_for_overflow_sub_gpu_0 =
-        (Torus **)malloc(num_blocks * sizeof(Torus *));
-
-    Torus *h_lut_indexes = (Torus *)malloc(num_blocks * sizeof(Torus));
-    Torus *h_scalar = (Torus *)malloc(num_blocks * sizeof(Torus));
-
-    // Extra indexes for the luts in first step
-    for (int nb = 1; nb <= num_blocks; nb++) {
-      first_indexes_for_overflow_sub_gpu_0[nb - 1] =
-          (Torus *)cuda_malloc_with_size_tracking_async(
-              nb * sizeof(Torus), streams.stream(0), streams.gpu_index(0),
-              size_tracker, allocate_gpu_memory);
-      for (int index = 0; index < nb; index++) {
-        uint32_t grouping_index = index / group_size;
-        bool is_in_first_grouping = (grouping_index == 0);
-        uint32_t index_in_grouping = index % group_size;
-        bool is_last_index = (index == (nb - 1));
-        if (is_last_index) {
-          if (nb == 1) {
-            h_lut_indexes[index] = 2 * group_size;
-          } else {
-            h_lut_indexes[index] = 2;
-          }
-        } else if (is_in_first_grouping) {
-          h_lut_indexes[index] = index_in_grouping;
-        } else {
-          h_lut_indexes[index] = index_in_grouping + group_size;
-        }
-      }
-      cuda_memcpy_with_size_tracking_async_to_gpu(
-          first_indexes_for_overflow_sub_gpu_0[nb - 1], h_lut_indexes,
-          nb * sizeof(Torus), streams.stream(0), streams.gpu_index(0),
-          allocate_gpu_memory);
-    }
-    // Extra indexes for the luts in second step
-    for (int nb = 1; nb <= num_blocks; nb++) {
-      second_indexes_for_overflow_sub_gpu_0[nb - 1] =
-          (Torus *)cuda_malloc_with_size_tracking_async(
-              nb * sizeof(Torus), streams.stream(0), streams.gpu_index(0),
-              size_tracker, allocate_gpu_memory);
-      scalars_for_overflow_sub_gpu_0[nb - 1] =
-          (Torus *)cuda_malloc_with_size_tracking_async(
-              nb * sizeof(Torus), streams.stream(0), streams.gpu_index(0),
-              size_tracker, allocate_gpu_memory);
-
-      for (int index = 0; index < nb; index++) {
-        uint32_t grouping_index = index / group_size;
-        bool is_in_first_grouping = (grouping_index == 0);
-        uint32_t index_in_grouping = index % group_size;
-
-        if (is_in_first_grouping) {
-          h_lut_indexes[index] = index_in_grouping;
-        } else if (index_in_grouping == (group_size - 1)) {
-          if (use_seq) {
-            int inner_index = (grouping_index - 1) % (group_size - 1);
-            h_lut_indexes[index] = inner_index + 2 * group_size;
-          } else {
-            h_lut_indexes[index] = 2 * group_size;
-          }
-        } else {
-          h_lut_indexes[index] = index_in_grouping + group_size;
-        }
-
-        bool may_have_its_padding_bit_set =
-            !is_in_first_grouping && (index_in_grouping == group_size - 1);
-
-        if (may_have_its_padding_bit_set) {
-          if (use_seq) {
-            h_scalar[index] = 1 << ((grouping_index - 1) % (group_size - 1));
-          } else {
-            h_scalar[index] = 1;
-          }
-        } else {
-          h_scalar[index] = 0;
-        }
-      }
-      cuda_memcpy_with_size_tracking_async_to_gpu(
-          second_indexes_for_overflow_sub_gpu_0[nb - 1], h_lut_indexes,
-          nb * sizeof(Torus), streams.stream(0), streams.gpu_index(0),
-          allocate_gpu_memory);
-      cuda_memcpy_with_size_tracking_async_to_gpu(
-          scalars_for_overflow_sub_gpu_0[nb - 1], h_scalar, nb * sizeof(Torus),
-          streams.stream(0), streams.gpu_index(0), allocate_gpu_memory);
-    }
-    cuda_synchronize_stream(streams.stream(0), streams.gpu_index(0));
-    free(h_lut_indexes);
-    free(h_scalar);
-  };
-
-  void release(CudaStreams streams) {
-
-    if (streams.count() < 4) {
-      PANIC("GPU count should be greater than 4 when using div_rem_2_2");
-    }
-    if (params.message_modulus != 4 || params.carry_modulus != 4) {
-      PANIC("Only message_modulus == 4 && carry_modulus == 4 parameters are "
-            "supported");
-    }
-
-    // release and delete integer ops memory objects
-    sub_and_propagate_mem->release(streams.get_ith(0));
-    shift_mem->release(streams.get_ith(1));
-    overflow_sub_mem_1->release(streams.get_ith(0));
-    overflow_sub_mem_2->release(streams.get_ith(1));
-    overflow_sub_mem_3->release(streams.get_ith(2));
-    comparison_buffer_1->release(streams.get_ith(0));
-    comparison_buffer_2->release(streams.get_ith(1));
-    comparison_buffer_3->release(streams.get_ith(2));
-    bitor_mem_1->release(streams.get_ith(0));
-    bitor_mem_2->release(streams.get_ith(1));
-    bitor_mem_3->release(streams.get_ith(2));
-
-    delete sub_and_propagate_mem;
-    sub_and_propagate_mem = nullptr;
-    delete shift_mem;
-    shift_mem = nullptr;
-    delete overflow_sub_mem_1;
-    overflow_sub_mem_1 = nullptr;
-    delete overflow_sub_mem_2;
-    overflow_sub_mem_2 = nullptr;
-    delete overflow_sub_mem_3;
-    overflow_sub_mem_3 = nullptr;
-    delete comparison_buffer_1;
-    comparison_buffer_1 = nullptr;
-    delete comparison_buffer_2;
-    comparison_buffer_2 = nullptr;
-    delete comparison_buffer_3;
-    comparison_buffer_3 = nullptr;
-    delete bitor_mem_1;
-    bitor_mem_1 = nullptr;
-    delete bitor_mem_2;
-    bitor_mem_2 = nullptr;
-    delete bitor_mem_3;
-    bitor_mem_3 = nullptr;
-
-    // release and delete lut objects
-    message_extract_lut_1->release(streams);
-    message_extract_lut_2->release(streams);
-    zero_out_if_not_1_lut_1->release(streams.get_ith(0));
-    zero_out_if_not_1_lut_2->release(streams.get_ith(3));
-    zero_out_if_not_2_lut_1->release(streams.get_ith(1));
-    zero_out_if_not_2_lut_2->release(streams.get_ith(2));
-    quotient_lut_1->release(streams.get_ith(2));
-    quotient_lut_2->release(streams.get_ith(1));
-    quotient_lut_3->release(streams.get_ith(0));
-
-    delete message_extract_lut_1;
-    message_extract_lut_1 = nullptr;
-    delete message_extract_lut_2;
-    message_extract_lut_2 = nullptr;
-    delete zero_out_if_not_1_lut_1;
-    zero_out_if_not_1_lut_1 = nullptr;
-    delete zero_out_if_not_1_lut_2;
-    zero_out_if_not_1_lut_2 = nullptr;
-    delete zero_out_if_not_2_lut_1;
-    zero_out_if_not_2_lut_1 = nullptr;
-    delete zero_out_if_not_2_lut_2;
-    zero_out_if_not_2_lut_2 = nullptr;
-    delete quotient_lut_1;
-    quotient_lut_1 = nullptr;
-    delete quotient_lut_2;
-    quotient_lut_2 = nullptr;
-    delete quotient_lut_3;
-    quotient_lut_3 = nullptr;
-
-    // release and delete temporary buffers
-    release_radix_ciphertext_async(streams.stream(2), streams.gpu_index(2), d1,
-                                   gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(1), streams.gpu_index(1), d2,
-                                   gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0), d3,
-                                   gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(2), streams.gpu_index(2),
-                                   low1, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(1), streams.gpu_index(1),
-                                   low2, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   low3, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(3), streams.gpu_index(3),
-                                   rem0, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(2), streams.gpu_index(2),
-                                   rem1, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(1), streams.gpu_index(1),
-                                   rem2, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   rem3, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(2), streams.gpu_index(2),
-                                   sub_result_3, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(1), streams.gpu_index(1),
-                                   sub_result_2, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   sub_result_1, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(2), streams.gpu_index(2),
-                                   sub_3_overflowed, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(1), streams.gpu_index(1),
-                                   sub_2_overflowed, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   sub_1_overflowed, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(3), streams.gpu_index(3),
-                                   tmp_gpu_3, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(2), streams.gpu_index(2),
-                                   tmp_gpu_2, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(1), streams.gpu_index(1),
-                                   tmp_gpu_1, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_gpu_0, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(1), streams.gpu_index(1),
-                                   divisor_gpu_1, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(2), streams.gpu_index(2),
-                                   divisor_gpu_2, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(1), streams.gpu_index(1),
-                                   remainder_gpu_1, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(2), streams.gpu_index(2),
-                                   remainder_gpu_2, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(3), streams.gpu_index(3),
-                                   remainder_gpu_3, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(2), streams.gpu_index(2),
-                                   comparison_blocks_3, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(1), streams.gpu_index(1),
-                                   comparison_blocks_2, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   comparison_blocks_1, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(2), streams.gpu_index(2),
-                                   cmp_3, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(1), streams.gpu_index(1),
-                                   cmp_2, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   cmp_1, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(3), streams.gpu_index(3), c0,
-                                   gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(2), streams.gpu_index(2), q1,
-                                   gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(1), streams.gpu_index(1), q2,
-                                   gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0), q3,
-                                   gpu_memory_allocated);
-
-    delete d1;
-    d1 = nullptr;
-    delete d2;
-    d2 = nullptr;
-    delete d3;
-    d3 = nullptr;
-    delete low1;
-    low1 = nullptr;
-    delete low2;
-    low2 = nullptr;
-    delete low3;
-    low3 = nullptr;
-    delete rem0;
-    rem0 = nullptr;
-    delete rem1;
-    rem1 = nullptr;
-    delete rem2;
-    rem2 = nullptr;
-    delete rem3;
-    rem3 = nullptr;
-    delete sub_result_1;
-    sub_result_1 = nullptr;
-    delete sub_result_2;
-    sub_result_2 = nullptr;
-    delete sub_result_3;
-    sub_result_3 = nullptr;
-    delete sub_1_overflowed;
-    sub_1_overflowed = nullptr;
-    delete sub_2_overflowed;
-    sub_2_overflowed = nullptr;
-    delete sub_3_overflowed;
-    sub_3_overflowed = nullptr;
-    delete tmp_gpu_0;
-    tmp_gpu_0 = nullptr;
-    delete tmp_gpu_1;
-    tmp_gpu_1 = nullptr;
-    delete tmp_gpu_2;
-    tmp_gpu_2 = nullptr;
-    delete tmp_gpu_3;
-    tmp_gpu_3 = nullptr;
-    delete divisor_gpu_1;
-    divisor_gpu_1 = nullptr;
-    delete divisor_gpu_2;
-    divisor_gpu_2 = nullptr;
-    delete remainder_gpu_1;
-    remainder_gpu_1 = nullptr;
-    delete remainder_gpu_2;
-    remainder_gpu_2 = nullptr;
-    delete remainder_gpu_3;
-    remainder_gpu_3 = nullptr;
-    delete comparison_blocks_1;
-    comparison_blocks_1 = nullptr;
-    delete comparison_blocks_2;
-    comparison_blocks_2 = nullptr;
-    delete comparison_blocks_3;
-    comparison_blocks_3 = nullptr;
-    delete cmp_1;
-    cmp_1 = nullptr;
-    delete cmp_2;
-    cmp_2 = nullptr;
-    delete cmp_3;
-    cmp_3 = nullptr;
-    delete c0;
-    c0 = nullptr;
-    delete q1;
-    q1 = nullptr;
-    delete q2;
-    q2 = nullptr;
-    delete q3;
-    q3 = nullptr;
-
-    for (int i = 0; i < max_indexes_to_erase; i++) {
-      cuda_drop_with_size_tracking_async(
-          first_indexes_for_overflow_sub_gpu_0[i], streams.stream(0),
-          streams.gpu_index(0), gpu_memory_allocated);
-      cuda_drop_with_size_tracking_async(
-          second_indexes_for_overflow_sub_gpu_0[i], streams.stream(0),
-          streams.gpu_index(0), gpu_memory_allocated);
-      cuda_drop_with_size_tracking_async(
-          scalars_for_overflow_sub_gpu_0[i], streams.stream(0),
-          streams.gpu_index(0), gpu_memory_allocated);
-      cuda_drop_with_size_tracking_async(
-          first_indexes_for_overflow_sub_gpu_1[i], streams.stream(1),
-          streams.gpu_index(1), gpu_memory_allocated);
-      cuda_drop_with_size_tracking_async(
-          second_indexes_for_overflow_sub_gpu_1[i], streams.stream(1),
-          streams.gpu_index(1), gpu_memory_allocated);
-      cuda_drop_with_size_tracking_async(
-          scalars_for_overflow_sub_gpu_1[i], streams.stream(1),
-          streams.gpu_index(1), gpu_memory_allocated);
-      cuda_drop_with_size_tracking_async(
-          first_indexes_for_overflow_sub_gpu_2[i], streams.stream(2),
-          streams.gpu_index(2), gpu_memory_allocated);
-      cuda_drop_with_size_tracking_async(
-          second_indexes_for_overflow_sub_gpu_2[i], streams.stream(2),
-          streams.gpu_index(2), gpu_memory_allocated);
-      cuda_drop_with_size_tracking_async(
-          scalars_for_overflow_sub_gpu_2[i], streams.stream(2),
-          streams.gpu_index(2), gpu_memory_allocated);
-    }
-
-    for (uint j = 0; j < 3; j++) {
-      cuda_synchronize_stream(streams.stream(j), streams.gpu_index(j));
-    }
-
-    free(first_indexes_for_overflow_sub_gpu_0);
-    free(second_indexes_for_overflow_sub_gpu_0);
-    free(scalars_for_overflow_sub_gpu_0);
-    free(first_indexes_for_overflow_sub_gpu_1);
-    free(second_indexes_for_overflow_sub_gpu_1);
-    free(scalars_for_overflow_sub_gpu_1);
-    free(first_indexes_for_overflow_sub_gpu_2);
-    free(second_indexes_for_overflow_sub_gpu_2);
-    free(scalars_for_overflow_sub_gpu_2);
-
-    cudaEventDestroy(create_indexes_done);
-
-    // release sub streams
-    sub_streams_1.release();
-  }
-};
-
-template <typename Torus> struct unsigned_int_div_rem_memory {
-  int_radix_params params;
-
-  // memory objects for other operations
-  int_logical_scalar_shift_buffer<Torus> *shift_mem_1;
-  int_logical_scalar_shift_buffer<Torus> *shift_mem_2;
-  int_borrow_prop_memory<Torus> *overflow_sub_mem;
-  int_comparison_buffer<Torus> *comparison_buffer;
-  unsigned_int_div_rem_2_2_memory<Torus> *div_rem_2_2_mem;
-
-  // lookup tables
-  int_radix_lut<Torus> **masking_luts_1;
-  int_radix_lut<Torus> **masking_luts_2;
-  int_radix_lut<Torus> *message_extract_lut_1;
-  int_radix_lut<Torus> *message_extract_lut_2;
-  int_radix_lut<Torus> **zero_out_if_overflow_did_not_happen;
-  int_radix_lut<Torus> **zero_out_if_overflow_happened;
-  int_radix_lut<Torus> **merge_overflow_flags_luts;
-
-  // sub streams
-  CudaStreams sub_streams_1;
-  CudaStreams sub_streams_2;
-  CudaStreams sub_streams_3;
-  CudaStreams sub_streams_4;
-
-  // temporary device buffers
-  CudaRadixCiphertextFFI *remainder1;
-  CudaRadixCiphertextFFI *remainder2;
-  CudaRadixCiphertextFFI *numerator_block_stack;
-  CudaRadixCiphertextFFI *numerator_block_1;
-  CudaRadixCiphertextFFI *tmp_radix;
-  CudaRadixCiphertextFFI *interesting_remainder1;
-  CudaRadixCiphertextFFI *interesting_remainder2;
-  CudaRadixCiphertextFFI *interesting_divisor;
-  CudaRadixCiphertextFFI *divisor_ms_blocks;
-  CudaRadixCiphertextFFI *new_remainder;
-  CudaRadixCiphertextFFI *subtraction_overflowed;
-  CudaRadixCiphertextFFI *did_not_overflow;
-  CudaRadixCiphertextFFI *overflow_sum;
-  CudaRadixCiphertextFFI *overflow_sum_radix;
-  CudaRadixCiphertextFFI *tmp_1;
-  CudaRadixCiphertextFFI *at_least_one_upper_block_is_non_zero;
-  CudaRadixCiphertextFFI *cleaned_merged_interesting_remainder;
-
-  Torus **first_indexes_for_overflow_sub;
-  Torus **second_indexes_for_overflow_sub;
-  Torus **scalars_for_overflow_sub;
-  uint32_t max_indexes_to_erase;
-  bool gpu_memory_allocated;
-
-  // allocate and initialize if needed, temporary arrays used to calculate
-  // cuda integer div_rem operation
-  void init_temporary_buffers(CudaStreams streams, uint32_t num_blocks,
-                              bool allocate_gpu_memory,
-                              uint64_t &size_tracker) {
-
-    // non boolean temporary arrays, with `num_blocks` blocks
-    remainder1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), remainder1, num_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    remainder2 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), remainder2, num_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    numerator_block_stack = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), numerator_block_stack,
-        num_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-    interesting_remainder2 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), interesting_remainder2,
-        num_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-    interesting_divisor = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), interesting_divisor,
-        num_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-    divisor_ms_blocks = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), divisor_ms_blocks, num_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    new_remainder = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), new_remainder, num_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    cleaned_merged_interesting_remainder = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0),
-        cleaned_merged_interesting_remainder, num_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    tmp_1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_1, num_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-    // temporary arrays used as stacks
-    tmp_radix = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_radix, num_blocks + 1,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    interesting_remainder1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), interesting_remainder1,
-        num_blocks + 1, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-    numerator_block_1 = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), numerator_block_1, 1,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-    // temporary arrays for boolean blocks
-    subtraction_overflowed = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), subtraction_overflowed, 1,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    did_not_overflow = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), did_not_overflow, 1,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    overflow_sum = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), overflow_sum, 1,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    overflow_sum_radix = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), overflow_sum_radix, num_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    at_least_one_upper_block_is_non_zero = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0),
-        at_least_one_upper_block_is_non_zero, 1, params.big_lwe_dimension,
-        size_tracker, allocate_gpu_memory);
-  }
-
-  // initialize lookup tables for div_rem operation
-  void init_lookup_tables(CudaStreams streams, uint32_t num_blocks,
-                          bool allocate_gpu_memory, uint64_t &size_tracker) {
-    uint32_t num_bits_in_message = 31 - __builtin_clz(params.message_modulus);
-
-    // create and generate masking_luts_1[] and masking_lut_2[]
-    // both of them are equal but because they are used in two different
-    // executions in parallel we need two different pbs_buffers.
-    masking_luts_1 = new int_radix_lut<Torus> *[params.message_modulus - 1];
-    masking_luts_2 = new int_radix_lut<Torus> *[params.message_modulus - 1];
-    for (int i = 0; i < params.message_modulus - 1; i++) {
-      uint32_t shifted_mask = i;
-      std::function<Torus(Torus)> lut_f_masking =
-          [shifted_mask](Torus x) -> Torus { return x & shifted_mask; };
-
-      masking_luts_1[i] = new int_radix_lut<Torus>(
-          streams, params, 1, 1, allocate_gpu_memory, size_tracker);
-      masking_luts_2[i] = new int_radix_lut<Torus>(
-          streams, params, 1, num_blocks, allocate_gpu_memory, size_tracker);
-
-      generate_device_accumulator<Torus>(
-          streams.stream(0), streams.gpu_index(0),
-          masking_luts_1[i]->get_lut(0, 0), masking_luts_1[i]->get_degree(0),
-          masking_luts_1[i]->get_max_degree(0), params.glwe_dimension,
-          params.polynomial_size, params.message_modulus, params.carry_modulus,
-          lut_f_masking, gpu_memory_allocated);
-      auto active_streams_1 = streams.active_gpu_subset(1);
-      masking_luts_1[i]->broadcast_lut(active_streams_1);
-
-      generate_device_accumulator<Torus>(
-          streams.stream(0), streams.gpu_index(0),
-          masking_luts_2[i]->get_lut(0, 0), masking_luts_2[i]->get_degree(0),
-          masking_luts_2[i]->get_max_degree(0), params.glwe_dimension,
-          params.polynomial_size, params.message_modulus, params.carry_modulus,
-          lut_f_masking, gpu_memory_allocated);
-      auto active_streams_2 = streams.active_gpu_subset(num_blocks);
-      masking_luts_2[i]->broadcast_lut(active_streams_2);
-    }
-
-    // create and generate message_extract_lut_1 and message_extract_lut_2
-    // both of them are equal but because they are used in two different
-    // executions in parallel we need two different pbs_buffers.
-    message_extract_lut_1 = new int_radix_lut<Torus>(
-        streams, params, 1, num_blocks, allocate_gpu_memory, size_tracker);
-    message_extract_lut_2 = new int_radix_lut<Torus>(
-        streams, params, 1, num_blocks, allocate_gpu_memory, size_tracker);
-
-    auto message_modulus = params.message_modulus;
-    auto lut_f_message_extract = [message_modulus](Torus x) -> Torus {
-      return x % message_modulus;
-    };
-
-    int_radix_lut<Torus> *luts[2] = {message_extract_lut_1,
-                                     message_extract_lut_2};
-    auto active_streams = streams.active_gpu_subset(num_blocks);
-    for (int j = 0; j < 2; j++) {
-      generate_device_accumulator<Torus>(
-          streams.stream(0), streams.gpu_index(0), luts[j]->get_lut(0, 0),
-          luts[j]->get_degree(0), luts[j]->get_max_degree(0),
-          params.glwe_dimension, params.polynomial_size, params.message_modulus,
-          params.carry_modulus, lut_f_message_extract, gpu_memory_allocated);
-      luts[j]->broadcast_lut(active_streams);
-    }
-
-    // Give name to closures to improve readability
-    auto overflow_happened = [](uint64_t overflow_sum) {
-      return overflow_sum != 0;
-    };
-    auto overflow_did_not_happen = [&overflow_happened](uint64_t overflow_sum) {
-      return !overflow_happened(overflow_sum);
-    };
-
-    // create and generate zero_out_if_overflow_did_not_happen
-    zero_out_if_overflow_did_not_happen = new int_radix_lut<Torus> *[2];
-    zero_out_if_overflow_did_not_happen[0] = new int_radix_lut<Torus>(
-        streams, params, 1, num_blocks, allocate_gpu_memory, size_tracker);
-    zero_out_if_overflow_did_not_happen[1] = new int_radix_lut<Torus>(
-        streams, params, 1, num_blocks, allocate_gpu_memory, size_tracker);
-
-    auto cur_lut_f = [&](Torus block, Torus overflow_sum) -> Torus {
-      if (overflow_did_not_happen(overflow_sum)) {
-        return 0;
-      } else {
-        return block;
-      }
-    };
-
-    generate_device_accumulator_bivariate_with_factor<Torus>(
-        streams.stream(0), streams.gpu_index(0),
-        zero_out_if_overflow_did_not_happen[0]->get_lut(0, 0),
-        zero_out_if_overflow_did_not_happen[0]->get_degree(0),
-        zero_out_if_overflow_did_not_happen[0]->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, cur_lut_f, params.message_modulus - 2,
-        gpu_memory_allocated);
-    zero_out_if_overflow_did_not_happen[0]->broadcast_lut(active_streams);
-    generate_device_accumulator_bivariate_with_factor<Torus>(
-        streams.stream(0), streams.gpu_index(0),
-        zero_out_if_overflow_did_not_happen[1]->get_lut(0, 0),
-        zero_out_if_overflow_did_not_happen[1]->get_degree(0),
-        zero_out_if_overflow_did_not_happen[1]->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, cur_lut_f, params.message_modulus - 1,
-        gpu_memory_allocated);
-    zero_out_if_overflow_did_not_happen[1]->broadcast_lut(active_streams);
-
-    // create and generate zero_out_if_overflow_happened
-    zero_out_if_overflow_happened = new int_radix_lut<Torus> *[2];
-    zero_out_if_overflow_happened[0] = new int_radix_lut<Torus>(
-        streams, params, 1, num_blocks, allocate_gpu_memory, size_tracker);
-    zero_out_if_overflow_happened[1] = new int_radix_lut<Torus>(
-        streams, params, 1, num_blocks, allocate_gpu_memory, size_tracker);
-
-    auto overflow_happened_f = [&](Torus block, Torus overflow_sum) -> Torus {
-      if (overflow_happened(overflow_sum)) {
-        return 0;
-      } else {
-        return block;
-      }
-    };
-
-    generate_device_accumulator_bivariate_with_factor<Torus>(
-        streams.stream(0), streams.gpu_index(0),
-        zero_out_if_overflow_happened[0]->get_lut(0, 0),
-        zero_out_if_overflow_happened[0]->get_degree(0),
-        zero_out_if_overflow_happened[0]->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, overflow_happened_f, params.message_modulus - 2,
-        gpu_memory_allocated);
-    zero_out_if_overflow_happened[0]->broadcast_lut(active_streams);
-    generate_device_accumulator_bivariate_with_factor<Torus>(
-        streams.stream(0), streams.gpu_index(0),
-        zero_out_if_overflow_happened[1]->get_lut(0, 0),
-        zero_out_if_overflow_happened[1]->get_degree(0),
-        zero_out_if_overflow_happened[1]->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, overflow_happened_f, params.message_modulus - 1,
-        gpu_memory_allocated);
-    zero_out_if_overflow_happened[1]->broadcast_lut(active_streams);
-
-    // merge_overflow_flags_luts
-    merge_overflow_flags_luts = new int_radix_lut<Torus> *[num_bits_in_message];
-    auto active_gpu_count_for_bits = streams.active_gpu_subset(1);
-    for (int i = 0; i < num_bits_in_message; i++) {
-      auto lut_f_bit = [i](Torus x, Torus y) -> Torus {
-        return (x == 0 && y == 0) << i;
-      };
-
-      merge_overflow_flags_luts[i] = new int_radix_lut<Torus>(
-          streams, params, 1, 1, allocate_gpu_memory, size_tracker);
-
-      generate_device_accumulator_bivariate<Torus>(
-          streams.stream(0), streams.gpu_index(0),
-          merge_overflow_flags_luts[i]->get_lut(0, 0),
-          merge_overflow_flags_luts[i]->get_degree(0),
-          merge_overflow_flags_luts[i]->get_max_degree(0),
-          params.glwe_dimension, params.polynomial_size, params.message_modulus,
-          params.carry_modulus, lut_f_bit, gpu_memory_allocated);
-      merge_overflow_flags_luts[i]->broadcast_lut(active_gpu_count_for_bits);
-    }
-  }
-
-  unsigned_int_div_rem_memory(CudaStreams streams, int_radix_params params,
-                              uint32_t num_blocks, bool allocate_gpu_memory,
-                              uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-    auto active_streams = streams.active_gpu_subset(2 * num_blocks);
-    this->params = params;
-
-    if (params.message_modulus == 4 && params.carry_modulus == 4 &&
-        streams.count() >= 4) {
-      div_rem_2_2_mem = new unsigned_int_div_rem_2_2_memory<Torus>(
-          streams, params, num_blocks, allocate_gpu_memory, size_tracker);
-      return;
-    }
-
-    shift_mem_1 = new int_logical_scalar_shift_buffer<Torus>(
-        streams, SHIFT_OR_ROTATE_TYPE::LEFT_SHIFT, params, 2 * num_blocks,
-        allocate_gpu_memory, size_tracker);
-
-    shift_mem_2 = new int_logical_scalar_shift_buffer<Torus>(
-        streams, SHIFT_OR_ROTATE_TYPE::LEFT_SHIFT, params, 2 * num_blocks,
-        allocate_gpu_memory, size_tracker);
-
-    uint32_t compute_overflow = 1;
-    overflow_sub_mem = new int_borrow_prop_memory<Torus>(
-        streams, params, num_blocks, compute_overflow, allocate_gpu_memory,
-        size_tracker);
-    uint32_t group_size = overflow_sub_mem->group_size;
-    bool use_seq = overflow_sub_mem->prop_simu_group_carries_mem
-                       ->use_sequential_algorithm_to_resolve_group_carries;
-    create_indexes_for_overflow_sub(streams, num_blocks, group_size, use_seq,
-                                    allocate_gpu_memory, size_tracker);
-
-    comparison_buffer = new int_comparison_buffer<Torus>(
-        streams, COMPARISON_TYPE::NE, params, num_blocks, false,
-        allocate_gpu_memory, size_tracker);
-
-    init_lookup_tables(streams, num_blocks, allocate_gpu_memory, size_tracker);
-    init_temporary_buffers(streams, num_blocks, allocate_gpu_memory,
-                           size_tracker);
-
-    sub_streams_1.create_on_same_gpus(active_streams);
-    sub_streams_2.create_on_same_gpus(active_streams);
-    sub_streams_3.create_on_same_gpus(active_streams);
-    sub_streams_4.create_on_same_gpus(active_streams);
-  }
-
-  void create_indexes_for_overflow_sub(CudaStreams streams, uint32_t num_blocks,
-                                       uint32_t group_size, bool use_seq,
-                                       bool allocate_gpu_memory,
-                                       uint64_t &size_tracker) {
-    max_indexes_to_erase = num_blocks;
-
-    first_indexes_for_overflow_sub =
-        (Torus **)malloc(num_blocks * sizeof(Torus *));
-    second_indexes_for_overflow_sub =
-        (Torus **)malloc(num_blocks * sizeof(Torus *));
-    scalars_for_overflow_sub = (Torus **)malloc(num_blocks * sizeof(Torus *));
-
-    Torus *h_lut_indexes = (Torus *)malloc(num_blocks * sizeof(Torus));
-    Torus *h_scalar = (Torus *)malloc(num_blocks * sizeof(Torus));
-
-    // Extra indexes for the luts in first step
-    for (int nb = 1; nb <= num_blocks; nb++) {
-      first_indexes_for_overflow_sub[nb - 1] =
-          (Torus *)cuda_malloc_with_size_tracking_async(
-              nb * sizeof(Torus), streams.stream(0), streams.gpu_index(0),
-              size_tracker, allocate_gpu_memory);
-      for (int index = 0; index < nb; index++) {
-        uint32_t grouping_index = index / group_size;
-        bool is_in_first_grouping = (grouping_index == 0);
-        uint32_t index_in_grouping = index % group_size;
-        bool is_last_index = (index == (nb - 1));
-        if (is_last_index) {
-          if (nb == 1) {
-            h_lut_indexes[index] = 2 * group_size;
-          } else {
-            h_lut_indexes[index] = 2;
-          }
-        } else if (is_in_first_grouping) {
-          h_lut_indexes[index] = index_in_grouping;
-        } else {
-          h_lut_indexes[index] = index_in_grouping + group_size;
-        }
-      }
-      cuda_memcpy_with_size_tracking_async_to_gpu(
-          first_indexes_for_overflow_sub[nb - 1], h_lut_indexes,
-          nb * sizeof(Torus), streams.stream(0), streams.gpu_index(0),
-          allocate_gpu_memory);
-    }
-    // Extra indexes for the luts in second step
-    for (int nb = 1; nb <= num_blocks; nb++) {
-      second_indexes_for_overflow_sub[nb - 1] =
-          (Torus *)cuda_malloc_with_size_tracking_async(
-              nb * sizeof(Torus), streams.stream(0), streams.gpu_index(0),
-              size_tracker, allocate_gpu_memory);
-      scalars_for_overflow_sub[nb - 1] =
-          (Torus *)cuda_malloc_with_size_tracking_async(
-              nb * sizeof(Torus), streams.stream(0), streams.gpu_index(0),
-              size_tracker, allocate_gpu_memory);
-
-      for (int index = 0; index < nb; index++) {
-        uint32_t grouping_index = index / group_size;
-        bool is_in_first_grouping = (grouping_index == 0);
-        uint32_t index_in_grouping = index % group_size;
-
-        if (is_in_first_grouping) {
-          h_lut_indexes[index] = index_in_grouping;
-        } else if (index_in_grouping == (group_size - 1)) {
-          if (use_seq) {
-            int inner_index = (grouping_index - 1) % (group_size - 1);
-            h_lut_indexes[index] = inner_index + 2 * group_size;
-          } else {
-            h_lut_indexes[index] = 2 * group_size;
-          }
-        } else {
-          h_lut_indexes[index] = index_in_grouping + group_size;
-        }
-
-        bool may_have_its_padding_bit_set =
-            !is_in_first_grouping && (index_in_grouping == group_size - 1);
-
-        if (may_have_its_padding_bit_set) {
-          if (use_seq) {
-            h_scalar[index] = 1 << ((grouping_index - 1) % (group_size - 1));
-          } else {
-            h_scalar[index] = 1;
-          }
-        } else {
-          h_scalar[index] = 0;
-        }
-      }
-      cuda_memcpy_with_size_tracking_async_to_gpu(
-          second_indexes_for_overflow_sub[nb - 1], h_lut_indexes,
-          nb * sizeof(Torus), streams.stream(0), streams.gpu_index(0),
-          allocate_gpu_memory);
-      cuda_memcpy_with_size_tracking_async_to_gpu(
-          scalars_for_overflow_sub[nb - 1], h_scalar, nb * sizeof(Torus),
-          streams.stream(0), streams.gpu_index(0), allocate_gpu_memory);
-    }
-    free(h_lut_indexes);
-    free(h_scalar);
-  };
-
-  void release(CudaStreams streams) {
-
-    if (params.message_modulus == 4 && params.carry_modulus == 4 &&
-        streams.count() >= 4) {
-      div_rem_2_2_mem->release(streams);
-      delete div_rem_2_2_mem;
-      return;
-    }
-
-    uint32_t num_bits_in_message = 31 - __builtin_clz(params.message_modulus);
-
-    // release and delete other operation memory objects
-    shift_mem_1->release(streams);
-    shift_mem_2->release(streams);
-    overflow_sub_mem->release(streams);
-    comparison_buffer->release(streams);
-    delete shift_mem_1;
-    delete shift_mem_2;
-    delete overflow_sub_mem;
-    delete comparison_buffer;
-
-    // drop temporary buffers
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   remainder1, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   remainder2, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   numerator_block_stack, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   numerator_block_1, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_radix, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   interesting_remainder1,
-                                   gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   interesting_remainder2,
-                                   gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   interesting_divisor, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   divisor_ms_blocks, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   new_remainder, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   subtraction_overflowed,
-                                   gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   did_not_overflow, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   overflow_sum, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   overflow_sum_radix, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_1, gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   at_least_one_upper_block_is_non_zero,
-                                   gpu_memory_allocated);
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   cleaned_merged_interesting_remainder,
-                                   gpu_memory_allocated);
-    // release and delete lookup tables
-
-    // masking_luts_1 and masking_luts_2
-    for (int i = 0; i < params.message_modulus - 1; i++) {
-      masking_luts_1[i]->release(streams);
-      masking_luts_2[i]->release(streams);
-
-      delete masking_luts_1[i];
-      delete masking_luts_2[i];
-    }
-    delete[] masking_luts_1;
-    delete[] masking_luts_2;
-
-    // message_extract_lut_1 and message_extract_lut_2
-    message_extract_lut_1->release(streams);
-    message_extract_lut_2->release(streams);
-
-    delete message_extract_lut_1;
-    delete message_extract_lut_2;
-
-    // zero_out_if_overflow_did_not_happen
-    zero_out_if_overflow_did_not_happen[0]->release(streams);
-    zero_out_if_overflow_did_not_happen[1]->release(streams);
-
-    delete zero_out_if_overflow_did_not_happen[0];
-    delete zero_out_if_overflow_did_not_happen[1];
-
-    delete[] zero_out_if_overflow_did_not_happen;
-
-    // zero_out_if_overflow_happened
-    zero_out_if_overflow_happened[0]->release(streams);
-    zero_out_if_overflow_happened[1]->release(streams);
-
-    delete zero_out_if_overflow_happened[0];
-    delete zero_out_if_overflow_happened[1];
-
-    delete[] zero_out_if_overflow_happened;
-
-    // merge_overflow_flags_luts
-    for (int i = 0; i < num_bits_in_message; i++) {
-      merge_overflow_flags_luts[i]->release(streams);
-
-      delete merge_overflow_flags_luts[i];
-    }
-    delete[] merge_overflow_flags_luts;
-
-    // release sub streams
-    cuda_synchronize_stream(streams.stream(0), streams.gpu_index(0));
-    sub_streams_1.release();
-    sub_streams_2.release();
-    sub_streams_3.release();
-    sub_streams_4.release();
-
-    // Delete temporary buffers
-    delete remainder1;
-    delete remainder2;
-    delete numerator_block_stack;
-    delete numerator_block_1;
-    delete tmp_radix;
-    delete interesting_remainder1;
-    delete interesting_remainder2;
-    delete interesting_divisor;
-    delete divisor_ms_blocks;
-    delete new_remainder;
-    delete subtraction_overflowed;
-    delete did_not_overflow;
-    delete overflow_sum;
-    delete overflow_sum_radix;
-    delete tmp_1;
-    delete at_least_one_upper_block_is_non_zero;
-    delete cleaned_merged_interesting_remainder;
-
-    for (int i = 0; i < max_indexes_to_erase; i++) {
-      cuda_drop_with_size_tracking_async(
-          first_indexes_for_overflow_sub[i], streams.stream(0),
-          streams.gpu_index(0), gpu_memory_allocated);
-      cuda_drop_with_size_tracking_async(
-          second_indexes_for_overflow_sub[i], streams.stream(0),
-          streams.gpu_index(0), gpu_memory_allocated);
-      cuda_drop_with_size_tracking_async(
-          scalars_for_overflow_sub[i], streams.stream(0), streams.gpu_index(0),
-          gpu_memory_allocated);
-    }
-    free(first_indexes_for_overflow_sub);
-    free(second_indexes_for_overflow_sub);
-    free(scalars_for_overflow_sub);
-  }
-};
-
-template <typename Torus> struct int_bitop_buffer {
-
-  int_radix_params params;
-  int_radix_lut<Torus> *lut;
-  BITOP_TYPE op;
-  bool gpu_memory_allocated;
-
-  int_bitop_buffer(CudaStreams streams, BITOP_TYPE op, int_radix_params params,
-                   uint32_t num_radix_blocks, bool allocate_gpu_memory,
-                   uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-    this->op = op;
-    this->params = params;
-    auto active_streams = streams.active_gpu_subset(num_radix_blocks);
-    switch (op) {
-    case BITAND:
-    case BITOR:
-    case BITXOR:
-      lut = new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                     allocate_gpu_memory, size_tracker);
-      {
-        auto lut_bivariate_f = [op](Torus lhs, Torus rhs) -> Torus {
-          if (op == BITOP_TYPE::BITAND) {
-            // AND
-            return lhs & rhs;
-          } else if (op == BITOP_TYPE::BITOR) {
-            // OR
-            return lhs | rhs;
-          } else {
-            // XOR
-            return lhs ^ rhs;
-          }
-        };
-
-        generate_device_accumulator_bivariate<Torus>(
-            streams.stream(0), streams.gpu_index(0), lut->get_lut(0, 0),
-            lut->get_degree(0), lut->get_max_degree(0), params.glwe_dimension,
-            params.polynomial_size, params.message_modulus,
-            params.carry_modulus, lut_bivariate_f, gpu_memory_allocated);
-        lut->broadcast_lut(active_streams);
-      }
-      break;
-    default:
-      // Scalar OP
-      lut = new int_radix_lut<Torus>(streams, params, params.message_modulus,
-                                     num_radix_blocks, allocate_gpu_memory,
-                                     size_tracker);
-
-      for (int i = 0; i < params.message_modulus; i++) {
-        auto rhs = i;
-
-        auto lut_univariate_scalar_f = [op, rhs](Torus x) -> Torus {
-          if (op == BITOP_TYPE::SCALAR_BITAND) {
-            // AND
-            return x & rhs;
-          } else if (op == BITOP_TYPE::SCALAR_BITOR) {
-            // OR
-            return x | rhs;
-          } else {
-            // XOR
-            return x ^ rhs;
-          }
-        };
-        generate_device_accumulator<Torus>(
-            streams.stream(0), streams.gpu_index(0), lut->get_lut(0, i),
-            lut->get_degree(i), lut->get_max_degree(i), params.glwe_dimension,
-            params.polynomial_size, params.message_modulus,
-            params.carry_modulus, lut_univariate_scalar_f,
-            gpu_memory_allocated);
-        lut->broadcast_lut(active_streams);
-      }
-    }
-  }
-
-  void release(CudaStreams streams) {
-    lut->release(streams);
-    delete lut;
-  }
-};
-
-template <typename Torus> struct int_scalar_mul_buffer {
-  int_radix_params params;
-  int_logical_scalar_shift_buffer<Torus> *logical_scalar_shift_buffer;
-  int_sum_ciphertexts_vec_memory<Torus> *sum_ciphertexts_vec_mem;
-  CudaRadixCiphertextFFI *preshifted_buffer;
-  CudaRadixCiphertextFFI *all_shifted_buffer;
-  int_sc_prop_memory<Torus> *sc_prop_mem;
-  bool anticipated_buffers_drop;
-  bool gpu_memory_allocated;
-  uint32_t num_ciphertext_bits;
-
-  int_scalar_mul_buffer(CudaStreams streams, int_radix_params params,
-                        uint32_t num_radix_blocks, uint32_t num_scalar_bits,
-                        bool allocate_gpu_memory, bool anticipated_buffer_drop,
-                        uint64_t &size_tracker) {
-    gpu_memory_allocated = allocate_gpu_memory;
-    this->params = params;
-    this->anticipated_buffers_drop = anticipated_buffer_drop;
-
-    uint32_t msg_bits = (uint32_t)std::log2(params.message_modulus);
-    num_ciphertext_bits = msg_bits * num_scalar_bits;
-
-    //// Contains all shifted values of lhs for shift in range (0..msg_bits)
-    //// The idea is that with these we can create all other shift that are
-    /// in / range (0..total_bits) for free (block rotation)
-    preshifted_buffer = new CudaRadixCiphertextFFI;
-    uint64_t anticipated_drop_mem = 0;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), preshifted_buffer,
-        msg_bits * num_radix_blocks, params.big_lwe_dimension,
-        anticipated_drop_mem, allocate_gpu_memory);
-
-    all_shifted_buffer = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), all_shifted_buffer,
-        num_ciphertext_bits * num_radix_blocks, params.big_lwe_dimension,
-        size_tracker, allocate_gpu_memory);
-
-    if (num_ciphertext_bits * num_radix_blocks >= num_radix_blocks + 2)
-      logical_scalar_shift_buffer = new int_logical_scalar_shift_buffer<Torus>(
-          streams, LEFT_SHIFT, params, num_radix_blocks, allocate_gpu_memory,
-          all_shifted_buffer, anticipated_drop_mem);
-    else
-      logical_scalar_shift_buffer = new int_logical_scalar_shift_buffer<Torus>(
-          streams, LEFT_SHIFT, params, num_radix_blocks, allocate_gpu_memory,
-          anticipated_drop_mem);
-
-    uint64_t last_step_mem = 0;
-    if (num_ciphertext_bits > 0) {
-      sum_ciphertexts_vec_mem = new int_sum_ciphertexts_vec_memory<Torus>(
-          streams, params, num_radix_blocks, num_ciphertext_bits, true,
-          allocate_gpu_memory, last_step_mem);
-    }
-    uint32_t requested_flag = outputFlag::FLAG_NONE;
-    sc_prop_mem = new int_sc_prop_memory<Torus>(
-        streams, params, num_radix_blocks, requested_flag, allocate_gpu_memory,
-        last_step_mem);
-    if (anticipated_buffer_drop) {
-      size_tracker += std::max(anticipated_drop_mem, last_step_mem);
-    } else {
-      size_tracker += anticipated_drop_mem + last_step_mem;
-    }
-  }
-
-  void release(CudaStreams streams) {
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   all_shifted_buffer, gpu_memory_allocated);
-    if (num_ciphertext_bits > 0) {
-      sum_ciphertexts_vec_mem->release(streams);
-      delete sum_ciphertexts_vec_mem;
-    }
-    sc_prop_mem->release(streams);
-    delete sc_prop_mem;
-    delete all_shifted_buffer;
-    release_buffers(streams);
-  }
-
-  void release_buffers(CudaStreams streams) {
-    if (preshifted_buffer) {
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     preshifted_buffer, gpu_memory_allocated);
-      delete preshifted_buffer;
-      preshifted_buffer = nullptr;
-    }
-
-    if (logical_scalar_shift_buffer) {
-      logical_scalar_shift_buffer->release(streams);
-      delete logical_scalar_shift_buffer;
-      logical_scalar_shift_buffer = nullptr;
-    }
-  }
-};
-
-template <typename Torus> struct int_abs_buffer {
-  int_radix_params params;
-
-  int_arithmetic_scalar_shift_buffer<Torus> *arithmetic_scalar_shift_mem;
-  int_sc_prop_memory<Torus> *scp_mem;
-  int_bitop_buffer<Torus> *bitxor_mem;
-
-  CudaRadixCiphertextFFI *mask;
-  bool allocate_gpu_memory;
-
-  int_abs_buffer(CudaStreams streams, int_radix_params params,
-                 uint32_t num_radix_blocks, bool allocate_gpu_memory,
-                 uint64_t &size_tracker) {
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-    arithmetic_scalar_shift_mem = new int_arithmetic_scalar_shift_buffer<Torus>(
-        streams, SHIFT_OR_ROTATE_TYPE::RIGHT_SHIFT, params, num_radix_blocks,
-        allocate_gpu_memory, size_tracker);
-    uint32_t requested_flag = outputFlag::FLAG_NONE;
-    scp_mem = new int_sc_prop_memory<Torus>(streams, params, num_radix_blocks,
-                                            requested_flag, allocate_gpu_memory,
-                                            size_tracker);
-    bitxor_mem = new int_bitop_buffer<Torus>(streams, BITOP_TYPE::BITXOR,
-                                             params, num_radix_blocks,
-                                             allocate_gpu_memory, size_tracker);
-
-    mask = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), mask, num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-  }
-
-  void release(CudaStreams streams) {
-    arithmetic_scalar_shift_mem->release(streams);
-    scp_mem->release(streams);
-    bitxor_mem->release(streams);
-
-    delete arithmetic_scalar_shift_mem;
-    delete scp_mem;
-    delete bitxor_mem;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   mask, this->allocate_gpu_memory);
-    cuda_synchronize_stream(streams.stream(0), streams.gpu_index(0));
-    delete mask;
-  }
-};
-
-template <typename Torus> struct int_div_rem_memory {
-  int_radix_params params;
-  CudaStreams active_streams;
-  bool is_signed;
-  // memory objects for other operations
-  unsigned_int_div_rem_memory<Torus> *unsigned_mem;
-  int_abs_buffer<Torus> *abs_mem_1;
-  int_abs_buffer<Torus> *abs_mem_2;
-  int_sc_prop_memory<Torus> *scp_mem_1;
-  int_sc_prop_memory<Torus> *scp_mem_2;
-  int_cmux_buffer<Torus> *cmux_quotient_mem;
-  int_cmux_buffer<Torus> *cmux_remainder_mem;
-
-  // lookup tables
-  int_radix_lut<Torus> *compare_signed_bits_lut;
-
-  // sub streams
-  CudaStreams sub_streams_1;
-  CudaStreams sub_streams_2;
-
-  // temporary device buffers
-  CudaRadixCiphertextFFI *positive_numerator;
-  CudaRadixCiphertextFFI *positive_divisor;
-  CudaRadixCiphertextFFI *sign_bits_are_different;
-  CudaRadixCiphertextFFI *negated_quotient;
-  CudaRadixCiphertextFFI *negated_remainder;
-  bool gpu_memory_allocated;
-
-  int_div_rem_memory(CudaStreams streams, int_radix_params params,
-                     bool is_signed, uint32_t num_blocks,
-                     bool allocate_gpu_memory, uint64_t &size_tracker) {
-
-    gpu_memory_allocated = allocate_gpu_memory;
-    this->active_streams = streams.active_gpu_subset(num_blocks);
-    this->params = params;
-    this->is_signed = is_signed;
-
-    unsigned_mem = new unsigned_int_div_rem_memory<Torus>(
-        streams, params, num_blocks, allocate_gpu_memory, size_tracker);
-
-    if (is_signed) {
-      Torus sign_bit_pos = 31 - __builtin_clz(params.message_modulus) - 1;
-
-      // init memory objects for other integer operations
-      abs_mem_1 = new int_abs_buffer<Torus>(streams, params, num_blocks,
-                                            allocate_gpu_memory, size_tracker);
-      abs_mem_2 = new int_abs_buffer<Torus>(streams, params, num_blocks,
-                                            allocate_gpu_memory, size_tracker);
-      uint32_t requested_flag = outputFlag::FLAG_NONE;
-      scp_mem_1 = new int_sc_prop_memory<Torus>(
-          streams, params, num_blocks, requested_flag, allocate_gpu_memory,
-          size_tracker);
-      scp_mem_2 = new int_sc_prop_memory<Torus>(
-          streams, params, num_blocks, requested_flag, allocate_gpu_memory,
-          size_tracker);
-
-      std::function<uint64_t(uint64_t)> quotient_predicate_lut_f =
-          [](uint64_t x) -> uint64_t { return x == 1; };
-      std::function<uint64_t(uint64_t)> remainder_predicate_lut_f =
-          [sign_bit_pos](uint64_t x) -> uint64_t {
-        return (x >> sign_bit_pos) == 1;
-      };
-
-      cmux_quotient_mem = new int_cmux_buffer<Torus>(
-          streams, quotient_predicate_lut_f, params, num_blocks,
-          allocate_gpu_memory, size_tracker);
-      cmux_remainder_mem = new int_cmux_buffer<Torus>(
-          streams, remainder_predicate_lut_f, params, num_blocks,
-          allocate_gpu_memory, size_tracker);
-      // init temporary memory buffers
-      positive_numerator = new CudaRadixCiphertextFFI;
-      create_zero_radix_ciphertext_async<Torus>(
-          streams.stream(0), streams.gpu_index(0), positive_numerator,
-          num_blocks, params.big_lwe_dimension, size_tracker,
-          allocate_gpu_memory);
-      positive_divisor = new CudaRadixCiphertextFFI;
-      create_zero_radix_ciphertext_async<Torus>(
-          streams.stream(0), streams.gpu_index(0), positive_divisor, num_blocks,
-          params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-      negated_quotient = new CudaRadixCiphertextFFI;
-      create_zero_radix_ciphertext_async<Torus>(
-          streams.stream(0), streams.gpu_index(0), negated_quotient, num_blocks,
-          params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-      negated_remainder = new CudaRadixCiphertextFFI;
-      create_zero_radix_ciphertext_async<Torus>(
-          streams.stream(0), streams.gpu_index(0), negated_remainder,
-          num_blocks, params.big_lwe_dimension, size_tracker,
-          allocate_gpu_memory);
-
-      // init boolean temporary buffers
-      sign_bits_are_different = new CudaRadixCiphertextFFI;
-      create_zero_radix_ciphertext_async<Torus>(
-          streams.stream(0), streams.gpu_index(0), sign_bits_are_different, 1,
-          params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-      // init sub streams
-      sub_streams_1.create_on_same_gpus(streams);
-      sub_streams_2.create_on_same_gpus(streams);
-
-      // init lookup tables
-      //  to extract and compare signed bits
-      auto f_compare_extracted_signed_bits = [sign_bit_pos](Torus x,
-                                                            Torus y) -> Torus {
-        Torus x_sign_bit = (x >> sign_bit_pos) & 1;
-        Torus y_sign_bit = (y >> sign_bit_pos) & 1;
-        return (Torus)(x_sign_bit != y_sign_bit);
-      };
-
-      compare_signed_bits_lut = new int_radix_lut<Torus>(
-          streams, params, 1, 1, allocate_gpu_memory, size_tracker);
-
-      generate_device_accumulator_bivariate<Torus>(
-          streams.stream(0), streams.gpu_index(0),
-          compare_signed_bits_lut->get_lut(0, 0),
-          compare_signed_bits_lut->get_degree(0),
-          compare_signed_bits_lut->get_max_degree(0), params.glwe_dimension,
-          params.polynomial_size, params.message_modulus, params.carry_modulus,
-          f_compare_extracted_signed_bits, gpu_memory_allocated);
-      auto active_gpu_count_cmp =
-          streams.active_gpu_subset(1); // only 1 block needed
-      compare_signed_bits_lut->broadcast_lut(active_gpu_count_cmp);
-    }
-  }
-
-  void release(CudaStreams streams) {
-    unsigned_mem->release(streams);
-    delete unsigned_mem;
-
-    if (is_signed) {
-      // release objects for other integer operations
-      abs_mem_1->release(streams);
-      abs_mem_2->release(streams);
-      scp_mem_1->release(streams);
-      scp_mem_2->release(streams);
-      cmux_quotient_mem->release(streams);
-      cmux_remainder_mem->release(streams);
-
-      delete abs_mem_1;
-      delete abs_mem_2;
-      delete scp_mem_1;
-      delete scp_mem_2;
-      delete cmux_quotient_mem;
-      delete cmux_remainder_mem;
-
-      // drop temporary buffers
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     positive_numerator, gpu_memory_allocated);
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     positive_divisor, gpu_memory_allocated);
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     sign_bits_are_different,
-                                     gpu_memory_allocated);
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     negated_quotient, gpu_memory_allocated);
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     negated_remainder, gpu_memory_allocated);
-      // release lookup tables
-      compare_signed_bits_lut->release(streams);
-      delete compare_signed_bits_lut;
-
-      // release sub streams
-      cuda_synchronize_stream(streams.stream(0), streams.gpu_index(0));
-      sub_streams_1.release();
-      sub_streams_2.release();
-
-      // delete temporary buffers
-      delete positive_numerator;
-      delete positive_divisor;
-      delete sign_bits_are_different;
-      delete negated_quotient;
-      delete negated_remainder;
-    }
-  }
-};
-
-template <typename Torus> struct int_scalar_mul_high_buffer {
-  int_radix_params params;
-  bool allocate_gpu_memory;
-
-  int_logical_scalar_shift_buffer<Torus> *logical_scalar_shift_mem;
-  int_scalar_mul_buffer<Torus> *scalar_mul_mem;
-
-  CudaRadixCiphertextFFI *tmp;
-
-  int_scalar_mul_high_buffer(CudaStreams streams, const int_radix_params params,
-                             uint32_t num_radix_blocks,
-                             uint32_t num_scalar_bits,
-                             const bool allocate_gpu_memory,
-                             uint64_t &size_tracker) {
-
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-
-    this->logical_scalar_shift_mem = new int_logical_scalar_shift_buffer<Torus>(
-        streams, RIGHT_SHIFT, params, 2 * num_radix_blocks, allocate_gpu_memory,
-        size_tracker);
-
-    this->scalar_mul_mem = new int_scalar_mul_buffer<Torus>(
-        streams, params, 2 * num_radix_blocks, num_scalar_bits,
-        allocate_gpu_memory, true, size_tracker);
-
-    this->tmp = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp, 2 * num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-  }
-
-  void release(CudaStreams streams) {
-
-    logical_scalar_shift_mem->release(streams);
-    delete logical_scalar_shift_mem;
-
-    scalar_mul_mem->release(streams);
-    delete scalar_mul_mem;
-    scalar_mul_mem = nullptr;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0), tmp,
-                                   allocate_gpu_memory);
-    delete tmp;
-  }
-};
-
-template <typename Torus> struct int_sub_and_propagate {
-  int_radix_params params;
-  bool allocate_gpu_memory;
-
-  CudaRadixCiphertextFFI *neg_rhs_array;
-
-  int_sc_prop_memory<Torus> *sc_prop_mem;
-
-  int_sub_and_propagate(CudaStreams streams, const int_radix_params params,
-                        uint32_t num_radix_blocks, uint32_t requested_flag_in,
-                        bool allocate_gpu_memory, uint64_t &size_tracker) {
-
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-
-    this->sc_prop_mem = new int_sc_prop_memory<Torus>(
-        streams, params, num_radix_blocks, requested_flag_in,
-        allocate_gpu_memory, size_tracker);
-
-    this->neg_rhs_array = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), neg_rhs_array,
-        num_radix_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-  }
-
-  void release(CudaStreams streams) {
-
-    sc_prop_mem->release(streams);
-    delete sc_prop_mem;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   neg_rhs_array, allocate_gpu_memory);
-    delete neg_rhs_array;
-  }
-};
-
-template <typename Torus> struct int_extend_radix_with_sign_msb_buffer {
-
-  int_radix_params params;
-  bool allocate_gpu_memory;
-
-  int_radix_lut<Torus> *lut = nullptr;
-
-  CudaRadixCiphertextFFI *last_block = nullptr;
-  CudaRadixCiphertextFFI *padding_block = nullptr;
-
-  int_extend_radix_with_sign_msb_buffer(CudaStreams streams,
-                                        const int_radix_params params,
-                                        uint32_t num_radix_blocks,
-                                        uint32_t num_additional_blocks,
-                                        const bool allocate_gpu_memory,
-                                        uint64_t &size_tracker) {
-
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-
-    if (num_additional_blocks != 0) {
-      this->lut = new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                           allocate_gpu_memory, size_tracker);
-
-      uint32_t bits_per_block = std::log2(params.message_modulus);
-      uint32_t msg_modulus = params.message_modulus;
-
-      generate_device_accumulator<Torus>(
-          streams.stream(0), streams.gpu_index(0), lut->get_lut(0, 0),
-          lut->get_degree(0), lut->get_max_degree(0), params.glwe_dimension,
-          params.polynomial_size, params.message_modulus, params.carry_modulus,
-          [msg_modulus, bits_per_block](Torus x) {
-            const auto xm = x % msg_modulus;
-            const auto sign_bit = (xm >> (bits_per_block - 1)) & 1;
-            return (Torus)((msg_modulus - 1) * sign_bit);
-          },
-          allocate_gpu_memory);
-
-      auto active_streams = streams.active_gpu_subset(num_radix_blocks);
-      lut->broadcast_lut(active_streams);
-
-      this->last_block = new CudaRadixCiphertextFFI;
-
-      create_zero_radix_ciphertext_async<Torus>(
-          streams.stream(0), streams.gpu_index(0), last_block, 1,
-          params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-      this->padding_block = new CudaRadixCiphertextFFI;
-
-      create_zero_radix_ciphertext_async<Torus>(
-          streams.stream(0), streams.gpu_index(0), padding_block, 1,
-          params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-    }
-  }
-
-  void release(CudaStreams streams) {
-
-    if (lut != nullptr) {
-      lut->release(streams);
-      delete lut;
-    }
-    if (last_block != nullptr) {
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     last_block, allocate_gpu_memory);
-      delete last_block;
-    }
-    if (padding_block != nullptr) {
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     padding_block, allocate_gpu_memory);
-      delete padding_block;
-    }
-  }
-};
-
-template <typename Torus> struct int_unsigned_scalar_div_mem {
-  int_radix_params params;
-  bool allocate_gpu_memory;
-
-  CudaRadixCiphertextFFI *tmp_ffi = nullptr;
-
-  int_logical_scalar_shift_buffer<Torus> *logical_scalar_shift_mem = nullptr;
-  int_scalar_mul_high_buffer<Torus> *scalar_mul_high_mem = nullptr;
-  int_sc_prop_memory<Torus> *scp_mem = nullptr;
-  int_sub_and_propagate<Torus> *sub_and_propagate_mem = nullptr;
-
-  int_unsigned_scalar_div_mem(CudaStreams streams,
-                              const int_radix_params params,
-                              uint32_t num_radix_blocks,
-                              const CudaScalarDivisorFFI *scalar_divisor_ffi,
-                              const bool allocate_gpu_memory,
-                              uint64_t &size_tracker) {
-
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-
-    if (!scalar_divisor_ffi->is_abs_divisor_one) {
-      if (scalar_divisor_ffi->is_divisor_pow2) {
-
-        logical_scalar_shift_mem = new int_logical_scalar_shift_buffer<Torus>(
-            streams, RIGHT_SHIFT, params, num_radix_blocks, allocate_gpu_memory,
-            size_tracker);
-
-      } else if (scalar_divisor_ffi->divisor_has_more_bits_than_numerator) {
-
-        tmp_ffi = new CudaRadixCiphertextFFI;
-        create_zero_radix_ciphertext_async<Torus>(
-            streams.stream(0), streams.gpu_index(0), tmp_ffi, num_radix_blocks,
-            params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-      } else if (scalar_divisor_ffi
-                     ->is_chosen_multiplier_geq_two_pow_numerator) {
-
-        logical_scalar_shift_mem = new int_logical_scalar_shift_buffer<Torus>(
-            streams, RIGHT_SHIFT, params, num_radix_blocks, allocate_gpu_memory,
-            size_tracker);
-        scalar_mul_high_mem = new int_scalar_mul_high_buffer<Torus>(
-            streams, params, num_radix_blocks, scalar_divisor_ffi->active_bits,
-            allocate_gpu_memory, size_tracker);
-        scp_mem = new int_sc_prop_memory<Torus>(
-            streams, params, num_radix_blocks, FLAG_NONE, allocate_gpu_memory,
-            size_tracker);
-        sub_and_propagate_mem = new int_sub_and_propagate<Torus>(
-            streams, params, num_radix_blocks, FLAG_NONE, allocate_gpu_memory,
-            size_tracker);
-        tmp_ffi = new CudaRadixCiphertextFFI;
-        create_zero_radix_ciphertext_async<Torus>(
-            streams.stream(0), streams.gpu_index(0), tmp_ffi, num_radix_blocks,
-            params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-      } else {
-
-        logical_scalar_shift_mem = new int_logical_scalar_shift_buffer<Torus>(
-            streams, RIGHT_SHIFT, params, num_radix_blocks, allocate_gpu_memory,
-            size_tracker);
-        scalar_mul_high_mem = new int_scalar_mul_high_buffer<Torus>(
-            streams, params, num_radix_blocks, scalar_divisor_ffi->active_bits,
-            allocate_gpu_memory, size_tracker);
-      }
-    }
-  }
-
-  void release(CudaStreams streams) {
-
-    if (logical_scalar_shift_mem != nullptr) {
-      logical_scalar_shift_mem->release(streams);
-      delete logical_scalar_shift_mem;
-    }
-    if (scalar_mul_high_mem != nullptr) {
-      scalar_mul_high_mem->release(streams);
-      delete scalar_mul_high_mem;
-    }
-    if (scp_mem != nullptr) {
-      scp_mem->release(streams);
-      delete scp_mem;
-    }
-    if (sub_and_propagate_mem != nullptr) {
-      sub_and_propagate_mem->release(streams);
-      delete sub_and_propagate_mem;
-    }
-    if (tmp_ffi != nullptr) {
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     tmp_ffi, allocate_gpu_memory);
-      delete tmp_ffi;
-    }
-  }
-};
-
-template <typename Torus> struct int_signed_scalar_mul_high_buffer {
-  int_radix_params params;
-  bool allocate_gpu_memory;
-
-  int_logical_scalar_shift_buffer<Torus> *logical_scalar_shift_mem;
-  int_scalar_mul_buffer<Torus> *scalar_mul_mem;
-  int_extend_radix_with_sign_msb_buffer<Torus> *extend_radix_mem;
-
-  CudaRadixCiphertextFFI *tmp;
-
-  int_signed_scalar_mul_high_buffer(CudaStreams streams,
-                                    const int_radix_params params,
-                                    uint32_t num_radix_blocks,
-                                    uint32_t num_scalar_bits,
-                                    const bool allocate_gpu_memory,
-                                    uint64_t &size_tracker) {
-
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-
-    this->logical_scalar_shift_mem = new int_logical_scalar_shift_buffer<Torus>(
-        streams, RIGHT_SHIFT, params, 2 * num_radix_blocks, allocate_gpu_memory,
-        size_tracker);
-
-    this->scalar_mul_mem = new int_scalar_mul_buffer<Torus>(
-        streams, params, 2 * num_radix_blocks, num_scalar_bits,
-        allocate_gpu_memory, true, size_tracker);
-
-    this->tmp = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp, 2 * num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-    this->extend_radix_mem = new int_extend_radix_with_sign_msb_buffer<Torus>(
-        streams, params, num_radix_blocks, num_radix_blocks,
-        allocate_gpu_memory, size_tracker);
-  }
-
-  void release(CudaStreams streams) {
-
-    logical_scalar_shift_mem->release(streams);
-    delete logical_scalar_shift_mem;
-
-    scalar_mul_mem->release(streams);
-    delete scalar_mul_mem;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0), tmp,
-                                   allocate_gpu_memory);
-    delete tmp;
-
-    extend_radix_mem->release(streams);
-    delete extend_radix_mem;
-  }
-};
-
-template <typename Torus> struct int_signed_scalar_div_mem {
-  int_radix_params params;
-  bool allocate_gpu_memory;
-
-  CudaRadixCiphertextFFI *tmp_ffi = nullptr;
-  CudaRadixCiphertextFFI *xsign_ffi = nullptr;
-
-  int_arithmetic_scalar_shift_buffer<Torus> *arithmetic_scalar_shift_mem =
-      nullptr;
-  int_logical_scalar_shift_buffer<Torus> *logical_scalar_shift_mem = nullptr;
-  int_signed_scalar_mul_high_buffer<Torus> *scalar_mul_high_mem = nullptr;
-  int_sc_prop_memory<Torus> *scp_mem = nullptr;
-  int_sub_and_propagate<Torus> *sub_and_propagate_mem = nullptr;
-
-  int_signed_scalar_div_mem(CudaStreams streams, const int_radix_params params,
-                            uint32_t num_radix_blocks,
-                            const CudaScalarDivisorFFI *scalar_divisor_ffi,
-                            const bool allocate_gpu_memory,
-                            uint64_t &size_tracker) {
-
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-
-    if (!scalar_divisor_ffi->chosen_multiplier_has_more_bits_than_numerator) {
-
-      if (scalar_divisor_ffi->is_abs_divisor_one &&
-          scalar_divisor_ffi->is_divisor_negative) {
-        tmp_ffi = new CudaRadixCiphertextFFI;
-
-        create_zero_radix_ciphertext_async<Torus>(
-            streams.stream(0), streams.gpu_index(0), tmp_ffi, num_radix_blocks,
-            params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-      } else if (!scalar_divisor_ffi->is_abs_divisor_one) {
-
-        tmp_ffi = new CudaRadixCiphertextFFI;
-        create_zero_radix_ciphertext_async<Torus>(
-            streams.stream(0), streams.gpu_index(0), tmp_ffi, num_radix_blocks,
-            params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-        arithmetic_scalar_shift_mem =
-            new int_arithmetic_scalar_shift_buffer<Torus>(
-                streams, RIGHT_SHIFT, params, num_radix_blocks,
-                allocate_gpu_memory, size_tracker);
-
-        if (scalar_divisor_ffi->is_divisor_pow2) {
-
-          logical_scalar_shift_mem = new int_logical_scalar_shift_buffer<Torus>(
-              streams, RIGHT_SHIFT, params, num_radix_blocks,
-              allocate_gpu_memory, size_tracker);
-          scp_mem = new int_sc_prop_memory<Torus>(
-              streams, params, num_radix_blocks, FLAG_NONE, allocate_gpu_memory,
-              size_tracker);
-
-        } else {
-
-          xsign_ffi = new CudaRadixCiphertextFFI;
-          create_zero_radix_ciphertext_async<Torus>(
-              streams.stream(0), streams.gpu_index(0), xsign_ffi,
-              num_radix_blocks, params.big_lwe_dimension, size_tracker,
-              allocate_gpu_memory);
-
-          scalar_mul_high_mem = new int_signed_scalar_mul_high_buffer<Torus>(
-              streams, params, num_radix_blocks,
-              scalar_divisor_ffi->active_bits, allocate_gpu_memory,
-              size_tracker);
-
-          sub_and_propagate_mem = new int_sub_and_propagate<Torus>(
-              streams, params, num_radix_blocks, FLAG_NONE, allocate_gpu_memory,
-              size_tracker);
-
-          if (scalar_divisor_ffi->is_chosen_multiplier_geq_two_pow_numerator) {
-            scp_mem = new int_sc_prop_memory<Torus>(
-                streams, params, num_radix_blocks, FLAG_NONE,
-                allocate_gpu_memory, size_tracker);
-          }
-        }
-      }
-    }
-  }
-
-  void release(CudaStreams streams) {
-
-    if (arithmetic_scalar_shift_mem != nullptr) {
-      arithmetic_scalar_shift_mem->release(streams);
-      delete arithmetic_scalar_shift_mem;
-    }
-    if (logical_scalar_shift_mem != nullptr) {
-      logical_scalar_shift_mem->release(streams);
-      delete logical_scalar_shift_mem;
-    }
-    if (scalar_mul_high_mem != nullptr) {
-      scalar_mul_high_mem->release(streams);
-      delete scalar_mul_high_mem;
-    }
-    if (scp_mem != nullptr) {
-      scp_mem->release(streams);
-      delete scp_mem;
-    }
-    if (sub_and_propagate_mem != nullptr) {
-      sub_and_propagate_mem->release(streams);
-      delete sub_and_propagate_mem;
-    }
-    if (tmp_ffi != nullptr) {
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     tmp_ffi, allocate_gpu_memory);
-      delete tmp_ffi;
-    }
-    if (xsign_ffi != nullptr) {
-      release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                     xsign_ffi, allocate_gpu_memory);
-      delete xsign_ffi;
-    }
-  }
-};
-
-template <typename Torus> struct int_unsigned_scalar_div_rem_buffer {
-  int_radix_params params;
-  bool allocate_gpu_memory;
-
-  CudaRadixCiphertextFFI *numerator_ct;
-
-  int_unsigned_scalar_div_mem<Torus> *unsigned_div_mem;
-  int_bitop_buffer<Torus> *bitop_mem = nullptr;
-  int_scalar_mul_buffer<Torus> *scalar_mul_mem = nullptr;
-  int_sub_and_propagate<Torus> *sub_and_propagate_mem = nullptr;
-
-  int_unsigned_scalar_div_rem_buffer(
-      CudaStreams streams, const int_radix_params params,
-      uint32_t num_radix_blocks, const CudaScalarDivisorFFI *scalar_divisor_ffi,
-      uint32_t const active_bits_divisor, const bool allocate_gpu_memory,
-      uint64_t &size_tracker) {
-
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-
-    this->numerator_ct = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), numerator_ct, num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-    this->unsigned_div_mem = new int_unsigned_scalar_div_mem<Torus>(
-        streams, params, num_radix_blocks, scalar_divisor_ffi,
-        allocate_gpu_memory, size_tracker);
-
-    if (scalar_divisor_ffi->is_divisor_pow2) {
-      this->bitop_mem = new int_bitop_buffer<Torus>(
-          streams, SCALAR_BITAND, params, num_radix_blocks, allocate_gpu_memory,
-          size_tracker);
-    } else {
-      if (!scalar_divisor_ffi->is_divisor_zero &&
-          !scalar_divisor_ffi->is_abs_divisor_one && num_radix_blocks != 0) {
-        this->scalar_mul_mem = new int_scalar_mul_buffer<Torus>(
-            streams, params, num_radix_blocks, active_bits_divisor,
-            allocate_gpu_memory, true, size_tracker);
-      }
-      this->sub_and_propagate_mem = new int_sub_and_propagate<Torus>(
-          streams, params, num_radix_blocks, FLAG_NONE, allocate_gpu_memory,
-          size_tracker);
-    }
-  }
-
-  void release(CudaStreams streams) {
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   numerator_ct, allocate_gpu_memory);
-    delete numerator_ct;
-
-    unsigned_div_mem->release(streams);
-    delete unsigned_div_mem;
-
-    if (bitop_mem != nullptr) {
-      bitop_mem->release(streams);
-      delete bitop_mem;
-    }
-    if (scalar_mul_mem != nullptr) {
-      scalar_mul_mem->release(streams);
-      delete scalar_mul_mem;
-    }
-    if (sub_and_propagate_mem != nullptr) {
-      sub_and_propagate_mem->release(streams);
-      delete sub_and_propagate_mem;
-    }
-  }
-};
-
-template <typename Torus> struct int_signed_scalar_div_rem_buffer {
-  int_radix_params params;
-  bool allocate_gpu_memory;
-
-  CudaRadixCiphertextFFI *numerator_ct;
-
-  int_signed_scalar_div_mem<Torus> *signed_div_mem;
-  int_logical_scalar_shift_buffer<Torus> *logical_scalar_shift_mem = nullptr;
-  int_scalar_mul_buffer<Torus> *scalar_mul_mem = nullptr;
-  int_sub_and_propagate<Torus> *sub_and_propagate_mem;
-  int_sc_prop_memory<Torus> *scp_mem;
-
-  int_signed_scalar_div_rem_buffer(
-      CudaStreams streams, const int_radix_params params,
-      uint32_t num_radix_blocks, const CudaScalarDivisorFFI *scalar_divisor_ffi,
-      uint32_t const active_bits_divisor, const bool allocate_gpu_memory,
-      uint64_t &size_tracker) {
-
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-
-    this->numerator_ct = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), numerator_ct, num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-    this->signed_div_mem = new int_signed_scalar_div_mem<Torus>(
-        streams, params, num_radix_blocks, scalar_divisor_ffi,
-        allocate_gpu_memory, size_tracker);
-
-    this->scp_mem = new int_sc_prop_memory<Torus>(
-        streams, params, num_radix_blocks, FLAG_NONE, allocate_gpu_memory,
-        size_tracker);
-
-    bool is_divisor_one = scalar_divisor_ffi->is_abs_divisor_one &&
-                          !scalar_divisor_ffi->is_divisor_negative;
-
-    if (!scalar_divisor_ffi->is_divisor_negative &&
-        scalar_divisor_ffi->is_divisor_pow2) {
-      this->logical_scalar_shift_mem =
-          new int_logical_scalar_shift_buffer<Torus>(
-              streams, LEFT_SHIFT, params, num_radix_blocks,
-              allocate_gpu_memory, size_tracker);
-
-    } else if (!scalar_divisor_ffi->is_divisor_zero && !is_divisor_one &&
-               num_radix_blocks != 0) {
-      this->scalar_mul_mem = new int_scalar_mul_buffer<Torus>(
-          streams, params, num_radix_blocks, active_bits_divisor,
-          allocate_gpu_memory, true, size_tracker);
-    }
-
-    this->sub_and_propagate_mem = new int_sub_and_propagate<Torus>(
-        streams, params, num_radix_blocks, FLAG_NONE, allocate_gpu_memory,
-        size_tracker);
-  }
-
-  void release(CudaStreams streams) {
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   numerator_ct, allocate_gpu_memory);
-    delete numerator_ct;
-    numerator_ct = nullptr;
-
-    signed_div_mem->release(streams);
-    delete signed_div_mem;
-    signed_div_mem = nullptr;
-
-    scp_mem->release(streams);
-    delete scp_mem;
-    scp_mem = nullptr;
-
-    if (logical_scalar_shift_mem != nullptr) {
-      logical_scalar_shift_mem->release(streams);
-      delete logical_scalar_shift_mem;
-      logical_scalar_shift_mem = nullptr;
-    }
-    if (scalar_mul_mem != nullptr) {
-      scalar_mul_mem->release(streams);
-      delete scalar_mul_mem;
-      scalar_mul_mem = nullptr;
-    }
-    sub_and_propagate_mem->release(streams);
-    delete sub_and_propagate_mem;
-    sub_and_propagate_mem = nullptr;
-  }
-};
-
-template <typename Torus> struct int_prepare_count_of_consecutive_bits_buffer {
-  int_radix_params params;
-  bool allocate_gpu_memory;
-
-  int_radix_lut<Torus> *univ_lut_mem;
-  int_radix_lut<Torus> *biv_lut_mem;
-
-  Direction direction;
-  BitValue bit_value;
-
-  CudaRadixCiphertextFFI *tmp_ct;
-
-  int_prepare_count_of_consecutive_bits_buffer(
-      CudaStreams streams, const int_radix_params params,
-      uint32_t num_radix_blocks, Direction direction, BitValue bit_value,
-      const bool allocate_gpu_memory, uint64_t &size_tracker) {
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-    this->direction = direction;
-    this->bit_value = bit_value;
-    auto active_streams = streams.active_gpu_subset(num_radix_blocks);
-    this->univ_lut_mem =
-        new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                 allocate_gpu_memory, size_tracker);
-    this->biv_lut_mem =
-        new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
-                                 allocate_gpu_memory, size_tracker);
-
-    const uint32_t num_bits = std::log2(this->params.message_modulus);
-
-    auto generate_uni_lut_lambda = [this, num_bits](Torus x) -> Torus {
-      x %= this->params.message_modulus;
-      uint64_t count = 0;
-
-      if (this->direction == Trailing) {
-        for (uint32_t i = 0; i < num_bits; ++i) {
-          if (((x >> i) & 1) != this->bit_value) {
-            break;
-          }
-          count++;
-        }
-      } else {
-        for (int32_t i = num_bits - 1; i >= 0; --i) {
-          if (((x >> i) & 1) != this->bit_value) {
-            break;
-          }
-          count++;
-        }
-      }
-      return count;
-    };
-
-    generate_device_accumulator<Torus>(
-        streams.stream(0), streams.gpu_index(0), univ_lut_mem->get_lut(0, 0),
-        univ_lut_mem->get_degree(0), univ_lut_mem->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, generate_uni_lut_lambda, allocate_gpu_memory);
-
-    univ_lut_mem->broadcast_lut(active_streams);
-
-    auto generate_bi_lut_lambda =
-        [num_bits](Torus block_num_bit_count,
-                   Torus more_significant_block_bit_count) -> Torus {
-      if (more_significant_block_bit_count == num_bits) {
-        return block_num_bit_count;
-      }
-      return 0;
-    };
-
-    generate_device_accumulator_bivariate<Torus>(
-        streams.stream(0), streams.gpu_index(0), biv_lut_mem->get_lut(0, 0),
-        biv_lut_mem->get_degree(0), biv_lut_mem->get_max_degree(0),
-        params.glwe_dimension, params.polynomial_size, params.message_modulus,
-        params.carry_modulus, generate_bi_lut_lambda, allocate_gpu_memory);
-
-    biv_lut_mem->broadcast_lut(active_streams);
-
-    this->tmp_ct = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), tmp_ct, num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-  }
-
-  void release(CudaStreams streams) {
-    univ_lut_mem->release(streams);
-    delete univ_lut_mem;
-
-    biv_lut_mem->release(streams);
-    delete biv_lut_mem;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   tmp_ct, allocate_gpu_memory);
-    delete tmp_ct;
-  }
-};
-
-template <typename Torus> struct int_count_of_consecutive_bits_buffer {
-  int_radix_params params;
-  bool allocate_gpu_memory;
-  uint32_t counter_num_blocks;
-
-  int_prepare_count_of_consecutive_bits_buffer<Torus> *prepare_mem = nullptr;
-  CudaRadixCiphertextFFI *ct_prepared = nullptr;
-
-  int_sum_ciphertexts_vec_memory<Torus> *sum_mem = nullptr;
-  int_sc_prop_memory<Torus> *propagate_mem = nullptr;
-  CudaRadixCiphertextFFI *cts = nullptr;
-
-  int_count_of_consecutive_bits_buffer(CudaStreams streams,
-                                       const int_radix_params params,
-                                       uint32_t num_radix_blocks,
-                                       uint32_t counter_num_blocks,
-                                       Direction direction, BitValue bit_value,
-                                       const bool allocate_gpu_memory,
-                                       uint64_t &size_tracker) {
-
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-    this->counter_num_blocks = counter_num_blocks;
-
-    this->ct_prepared = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), ct_prepared, num_radix_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-    this->prepare_mem = new int_prepare_count_of_consecutive_bits_buffer<Torus>(
-        streams, params, num_radix_blocks, direction, bit_value,
-        allocate_gpu_memory, size_tracker);
-
-    this->cts = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), cts,
-        counter_num_blocks * num_radix_blocks, params.big_lwe_dimension,
-        size_tracker, allocate_gpu_memory);
-
-    this->sum_mem = new int_sum_ciphertexts_vec_memory<Torus>(
-        streams, params, counter_num_blocks, num_radix_blocks, true,
-        allocate_gpu_memory, size_tracker);
-
-    this->propagate_mem = new int_sc_prop_memory<Torus>(
-        streams, params, counter_num_blocks, FLAG_NONE, allocate_gpu_memory,
-        size_tracker);
-  }
-
-  void release(CudaStreams streams) {
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   ct_prepared, allocate_gpu_memory);
-    delete ct_prepared;
-    ct_prepared = nullptr;
-
-    prepare_mem->release(streams);
-    delete prepare_mem;
-    prepare_mem = nullptr;
-
-    sum_mem->release(streams);
-    delete sum_mem;
-    sum_mem = nullptr;
-
-    propagate_mem->release(streams);
-    delete propagate_mem;
-    propagate_mem = nullptr;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0), cts,
-                                   allocate_gpu_memory);
-    delete cts;
-    cts = nullptr;
-  }
-};
-
-template <typename Torus> struct int_grouped_oprf_memory {
-  int_radix_params params;
-  bool allocate_gpu_memory;
-
-  int_radix_lut<Torus> *luts;
-  CudaRadixCiphertextFFI *plaintext_corrections;
-  Torus *h_lut_indexes;
-
-  // with message_bits_per_block == ilog2(msg_modulus) from crypto params
-  int_grouped_oprf_memory(CudaStreams streams, int_radix_params params,
-                          uint32_t num_blocks_to_process,
-                          uint32_t message_bits_per_block,
-                          uint64_t total_random_bits, bool allocate_gpu_memory,
-                          uint64_t &size_tracker) {
-
-    uint32_t calculated_active_blocks =
-        total_random_bits == 0
-            ? 0
-            : (total_random_bits + message_bits_per_block - 1) /
-                  message_bits_per_block;
-    if (num_blocks_to_process != calculated_active_blocks) {
-      PANIC(
-          "num_blocks_to_process should be equal to calculated_active_blocks");
-    }
-
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-
-    this->luts = new int_radix_lut<Torus>(
-        streams, params, message_bits_per_block, num_blocks_to_process,
-        allocate_gpu_memory, size_tracker);
-
-    this->plaintext_corrections = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), this->plaintext_corrections,
-        num_blocks_to_process, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    uint64_t message_modulus_log2 = (uint64_t)std::log2(params.message_modulus);
-    if (message_modulus_log2 != message_bits_per_block) {
-      PANIC("message_modulus_log2 should be equal to message_bits_per_block");
-    }
-    uint64_t carry_modulus_log2 = (uint64_t)std::log2(params.carry_modulus);
-    uint64_t full_bits_count = 1 + carry_modulus_log2 + message_modulus_log2;
-    uint64_t delta = 1ULL << (64 - full_bits_count);
-    size_t lwe_size = params.big_lwe_dimension + 1;
-
-    // Pre-generate all possible LUTs.
-    //
-    for (uint32_t random_bit = 1; random_bit <= message_bits_per_block;
-         ++random_bit) {
-      uint64_t p = 1ULL << random_bit;
-      uint64_t poly_delta =
-          2 * static_cast<uint64_t>(params.polynomial_size) / p;
-
-      if (2 * static_cast<uint64_t>(params.polynomial_size) < p) {
-        PANIC("2 * static_cast<uint64_t>(params.polynomial_size) should be "
-              "smaller than p");
-      }
-
-      auto lut_f = [poly_delta, delta](uint32_t x) -> Torus {
-        return (2 * (x / poly_delta) + 1) * delta / 2;
-      };
-
-      uint64_t degree = 0;
-      uint32_t lut_index = random_bit - 1;
-      generate_device_accumulator_no_encoding<Torus>(
-          streams.stream(0), streams.gpu_index(0), luts->get_lut(0, lut_index),
-          degree, params.message_modulus, params.carry_modulus,
-          params.glwe_dimension, params.polynomial_size, lut_f,
-          allocate_gpu_memory);
-      // In  OPRF the degree is hard set to p - 1 instead of the LUT degree
-      degree = p - 1;
-      *luts->get_degree(lut_index) = degree;
-    }
-
-    // For each block, this loop determines the exact number of bits to generate
-    // (handling both bounded and unbounded cases), which pre-computed LUT to
-    // use, and the final plaintext correction to add.
-    //
-    Torus *h_corrections =
-        (Torus *)calloc(num_blocks_to_process * lwe_size, sizeof(Torus));
-    this->h_lut_indexes = (Torus *)calloc(num_blocks_to_process, sizeof(Torus));
-
-    uint64_t bits_processed = 0;
-    for (uint32_t i = 0; i < num_blocks_to_process; ++i) {
-
-      if (total_random_bits <= bits_processed) {
-        PANIC("total_random_bits should be greater than bits_processed");
-      }
-      uint64_t bits_remaining = total_random_bits - bits_processed;
-      uint32_t bits_for_this_block =
-          std::min((uint64_t)message_bits_per_block, bits_remaining);
-
-      uint64_t p = 1ULL << bits_for_this_block;
-      Torus plaintext_to_add = (p - 1) * delta / 2;
-
-      h_corrections[i * lwe_size + params.big_lwe_dimension] = plaintext_to_add;
-      if (bits_for_this_block < 1) {
-        PANIC("bits_for_this_block should be greater than 1");
-      }
-      this->h_lut_indexes[i] = bits_for_this_block - 1;
-
-      bits_processed += bits_for_this_block;
-    }
-
-    // h_corrections contains num_blocks lwes of dimension big_lwe_dim
-    // of which num_blocks_to_process lwes have a body that his set
-    // to a correction and all others to 0.
-    // All lwes in h_corrections have a mask equal to 0.
-    // Copy the prepared plaintext corrections to the GPU.
-    cuda_memcpy_with_size_tracking_async_to_gpu(
-        this->plaintext_corrections->ptr, h_corrections,
-        num_blocks_to_process * lwe_size * sizeof(Torus), streams.stream(0),
-        streams.gpu_index(0), allocate_gpu_memory);
-
-    // Copy the prepared LUT indexes to the GPU 0, before broadcast to all other
-    // GPUs.
-    cuda_memcpy_with_size_tracking_async_to_gpu(
-        luts->get_lut_indexes(0, 0), this->h_lut_indexes,
-        num_blocks_to_process * sizeof(Torus), streams.stream(0),
-        streams.gpu_index(0), allocate_gpu_memory);
-    auto active_streams = streams.active_gpu_subset(num_blocks_to_process);
-    luts->broadcast_lut(active_streams);
-
-    free(h_corrections);
-  }
-
-  void release(CudaStreams streams) {
-    this->luts->release(streams);
-    delete this->luts;
-    this->luts = nullptr;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   this->plaintext_corrections,
-                                   this->allocate_gpu_memory);
-    delete this->plaintext_corrections;
-    this->plaintext_corrections = nullptr;
-
-    free(this->h_lut_indexes);
-    this->h_lut_indexes = nullptr;
-  }
-};
-
-template <typename Torus> struct int_ilog2_buffer {
-  int_radix_params params;
-  bool allocate_gpu_memory;
-  uint32_t input_num_blocks;
-  uint32_t counter_num_blocks;
-  uint32_t num_bits_in_ciphertext;
-
-  int_prepare_count_of_consecutive_bits_buffer<Torus> *prepare_mem;
-  int_sum_ciphertexts_vec_memory<Torus> *sum_mem;
-  int_fullprop_buffer<Torus> *final_propagate_mem;
-
-  CudaRadixCiphertextFFI *ct_in_buffer;
-  CudaRadixCiphertextFFI *sum_input_cts;
-  CudaRadixCiphertextFFI *sum_output_not_propagated;
-  CudaRadixCiphertextFFI *message_blocks_not;
-  CudaRadixCiphertextFFI *carry_blocks_not;
-  CudaRadixCiphertextFFI *rotated_carry_blocks;
-
-  int_radix_lut<Torus> *lut_message_not;
-  int_radix_lut<Torus> *lut_carry_not;
-
-  int_ilog2_buffer(CudaStreams streams, const int_radix_params params,
-                   uint32_t input_num_blocks, uint32_t counter_num_blocks,
-                   uint32_t num_bits_in_ciphertext,
-                   const bool allocate_gpu_memory, uint64_t &size_tracker) {
-
-    this->params = params;
-    this->allocate_gpu_memory = allocate_gpu_memory;
-    this->input_num_blocks = input_num_blocks;
-    this->counter_num_blocks = counter_num_blocks;
-    this->num_bits_in_ciphertext = num_bits_in_ciphertext;
-
-    this->ct_in_buffer = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), this->ct_in_buffer,
-        input_num_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    this->prepare_mem = new int_prepare_count_of_consecutive_bits_buffer<Torus>(
-        streams, params, input_num_blocks, Leading, Zero, allocate_gpu_memory,
-        size_tracker);
-
-    uint32_t sum_input_total_blocks =
-        (input_num_blocks + 1) * counter_num_blocks;
-    this->sum_input_cts = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), this->sum_input_cts,
-        sum_input_total_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    this->sum_mem = new int_sum_ciphertexts_vec_memory<Torus>(
-        streams, params, counter_num_blocks, input_num_blocks + 1, false,
-        allocate_gpu_memory, size_tracker);
-
-    this->sum_output_not_propagated = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0),
-        this->sum_output_not_propagated, counter_num_blocks,
-        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
-
-    this->lut_message_not =
-        new int_radix_lut<Torus>(streams, params, 1, counter_num_blocks,
-                                 allocate_gpu_memory, size_tracker);
-    std::function<Torus(Torus)> lut_message_lambda =
-        [this](uint64_t x) -> uint64_t {
-      uint64_t message = x % this->params.message_modulus;
-      return (~message) % this->params.message_modulus;
-    };
-    generate_device_accumulator(streams.stream(0), streams.gpu_index(0),
-                                this->lut_message_not->get_lut(0, 0),
-                                this->lut_message_not->get_degree(0),
-                                this->lut_message_not->get_max_degree(0),
-                                params.glwe_dimension, params.polynomial_size,
-                                params.message_modulus, params.carry_modulus,
-                                lut_message_lambda, allocate_gpu_memory);
-    auto active_streams = streams.active_gpu_subset(counter_num_blocks);
-    lut_message_not->broadcast_lut(active_streams);
-
-    this->lut_carry_not =
-        new int_radix_lut<Torus>(streams, params, 1, counter_num_blocks,
-                                 allocate_gpu_memory, size_tracker);
-    std::function<Torus(Torus)> lut_carry_lambda =
-        [this](uint64_t x) -> uint64_t {
-      uint64_t carry = x / this->params.message_modulus;
-      return (~carry) % this->params.message_modulus;
-    };
-    generate_device_accumulator(
-        streams.stream(0), streams.gpu_index(0),
-        this->lut_carry_not->get_lut(0, 0), this->lut_carry_not->get_degree(0),
-        this->lut_carry_not->get_max_degree(0), params.glwe_dimension,
-        params.polynomial_size, params.message_modulus, params.carry_modulus,
-        lut_carry_lambda, allocate_gpu_memory);
-    lut_carry_not->broadcast_lut(active_streams);
-
-    this->message_blocks_not = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), this->message_blocks_not,
-        counter_num_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    this->carry_blocks_not = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), this->carry_blocks_not,
-        counter_num_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    this->rotated_carry_blocks = new CudaRadixCiphertextFFI;
-    create_zero_radix_ciphertext_async<Torus>(
-        streams.stream(0), streams.gpu_index(0), this->rotated_carry_blocks,
-        counter_num_blocks, params.big_lwe_dimension, size_tracker,
-        allocate_gpu_memory);
-
-    this->final_propagate_mem = new int_fullprop_buffer<Torus>(
-        streams, params, allocate_gpu_memory, size_tracker);
-  }
-
-  void release(CudaStreams streams) {
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   this->ct_in_buffer, allocate_gpu_memory);
-    delete this->ct_in_buffer;
-    this->ct_in_buffer = nullptr;
-
-    this->prepare_mem->release(streams);
-    delete this->prepare_mem;
-    this->prepare_mem = nullptr;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   this->sum_input_cts, allocate_gpu_memory);
-    delete this->sum_input_cts;
-    this->sum_input_cts = nullptr;
-
-    this->sum_mem->release(streams);
-    delete this->sum_mem;
-    this->sum_mem = nullptr;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   this->sum_output_not_propagated,
-                                   allocate_gpu_memory);
-    delete this->sum_output_not_propagated;
-    this->sum_output_not_propagated = nullptr;
-
-    this->lut_message_not->release(streams);
-    delete this->lut_message_not;
-    this->lut_message_not = nullptr;
-
-    this->lut_carry_not->release(streams);
-    delete this->lut_carry_not;
-    this->lut_carry_not = nullptr;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   this->message_blocks_not,
-                                   allocate_gpu_memory);
-    delete this->message_blocks_not;
-    this->message_blocks_not = nullptr;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   this->carry_blocks_not, allocate_gpu_memory);
-    delete this->carry_blocks_not;
-    this->carry_blocks_not = nullptr;
-
-    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
-                                   this->rotated_carry_blocks,
-                                   allocate_gpu_memory);
-    delete this->rotated_carry_blocks;
-    this->rotated_carry_blocks = nullptr;
-
-    this->final_propagate_mem->release(streams);
-    delete this->final_propagate_mem;
-    this->final_propagate_mem = nullptr;
-  }
-};
-
-void update_degrees_after_bitand(uint64_t *output_degrees,
-                                 uint64_t *lwe_array_1_degrees,
-                                 uint64_t *lwe_array_2_degrees,
-                                 uint32_t num_radix_blocks);
-void update_degrees_after_bitor(uint64_t *output_degrees,
-                                uint64_t *lwe_array_1_degrees,
-                                uint64_t *lwe_array_2_degrees,
-                                uint32_t num_radix_blocks);
-void update_degrees_after_bitxor(uint64_t *output_degrees,
-                                 uint64_t *lwe_array_1_degrees,
-                                 uint64_t *lwe_array_2_degrees,
-                                 uint32_t num_radix_blocks);
-void update_degrees_after_scalar_bitand(uint64_t *output_degrees,
-                                        uint64_t const *clear_degrees,
-                                        uint64_t const *input_degrees,
-                                        uint32_t num_clear_blocks);
-void update_degrees_after_scalar_bitor(uint64_t *output_degrees,
-                                       uint64_t const *clear_degrees,
-                                       uint64_t const *input_degrees,
-                                       uint32_t num_clear_blocks);
-void update_degrees_after_scalar_bitxor(uint64_t *output_degrees,
-                                        uint64_t const *clear_degrees,
-                                        uint64_t const *input_degrees,
-                                        uint32_t num_clear_blocks);
 std::pair<bool, bool> get_invert_flags(COMPARISON_TYPE compare);
 void reverseArray(uint64_t arr[], size_t n);
 #endif // CUDA_INTEGER_UTILITIES_H
