+++ backends/tfhe-cuda-backend/cuda/include/integer/shift_and_rotate.h
@@ -1,0 +1,169 @@
+#pragma once
+#include "integer_utilities.h"
+
+template <typename Torus> struct int_shift_and_rotate_buffer {
+  int_radix_params params;
+  SHIFT_OR_ROTATE_TYPE shift_type;
+  bool is_signed;
+
+  CudaRadixCiphertextFFI *tmp_bits;
+  CudaRadixCiphertextFFI *tmp_shift_bits;
+  CudaRadixCiphertextFFI *tmp_rotated;
+  CudaRadixCiphertextFFI *tmp_input_bits_a;
+  CudaRadixCiphertextFFI *tmp_input_bits_b;
+  CudaRadixCiphertextFFI *tmp_mux_inputs;
+
+  int_bit_extract_luts_buffer<Torus> *bit_extract_luts;
+  int_bit_extract_luts_buffer<Torus> *bit_extract_luts_with_offset_2;
+  int_radix_lut<Torus> *mux_lut;
+  int_radix_lut<Torus> *cleaning_lut;
+
+  Torus offset;
+  bool gpu_memory_allocated;
+
+  int_shift_and_rotate_buffer(CudaStreams streams,
+                              SHIFT_OR_ROTATE_TYPE shift_type, bool is_signed,
+                              int_radix_params params,
+                              uint32_t num_radix_blocks,
+                              bool allocate_gpu_memory,
+                              uint64_t &size_tracker) {
+    this->shift_type = shift_type;
+    this->is_signed = is_signed;
+    this->params = params;
+    gpu_memory_allocated = allocate_gpu_memory;
+
+    uint32_t bits_per_block = std::log2(params.message_modulus);
+    uint32_t total_nb_bits =
+        std::log2(params.message_modulus) * num_radix_blocks;
+    uint32_t max_num_bits_that_tell_shift = std::log2(total_nb_bits);
+
+    auto is_power_of_two = [](uint32_t n) {
+      return (n > 0) && ((n & (n - 1)) == 0);
+    };
+
+    if (!is_power_of_two(total_nb_bits))
+      max_num_bits_that_tell_shift += 1;
+
+    offset = (shift_type == LEFT_SHIFT ? 0 : total_nb_bits);
+
+    bit_extract_luts = new int_bit_extract_luts_buffer<Torus>(
+        streams, params, bits_per_block, num_radix_blocks, allocate_gpu_memory,
+        size_tracker);
+    bit_extract_luts_with_offset_2 = new int_bit_extract_luts_buffer<Torus>(
+        streams, params, bits_per_block, 2, num_radix_blocks,
+        allocate_gpu_memory, size_tracker);
+
+    mux_lut = new int_radix_lut<Torus>(streams, params, 1,
+                                       bits_per_block * num_radix_blocks,
+                                       allocate_gpu_memory, size_tracker);
+    cleaning_lut =
+        new int_radix_lut<Torus>(streams, params, 1, num_radix_blocks,
+                                 allocate_gpu_memory, size_tracker);
+
+    tmp_bits = new CudaRadixCiphertextFFI;
+    create_zero_radix_ciphertext_async<Torus>(
+        streams.stream(0), streams.gpu_index(0), tmp_bits,
+        bits_per_block * num_radix_blocks, params.big_lwe_dimension,
+        size_tracker, allocate_gpu_memory);
+
+    tmp_shift_bits = new CudaRadixCiphertextFFI;
+    create_zero_radix_ciphertext_async<Torus>(
+        streams.stream(0), streams.gpu_index(0), tmp_shift_bits,
+        max_num_bits_that_tell_shift * num_radix_blocks,
+        params.big_lwe_dimension, size_tracker, allocate_gpu_memory);
+
+    tmp_rotated = new CudaRadixCiphertextFFI;
+    create_zero_radix_ciphertext_async<Torus>(
+        streams.stream(0), streams.gpu_index(0), tmp_rotated,
+        bits_per_block * num_radix_blocks, params.big_lwe_dimension,
+        size_tracker, allocate_gpu_memory);
+
+    tmp_input_bits_a = new CudaRadixCiphertextFFI;
+    create_zero_radix_ciphertext_async<Torus>(
+        streams.stream(0), streams.gpu_index(0), tmp_input_bits_a,
+        bits_per_block * num_radix_blocks, params.big_lwe_dimension,
+        size_tracker, allocate_gpu_memory);
+
+    tmp_input_bits_b = new CudaRadixCiphertextFFI;
+    create_zero_radix_ciphertext_async<Torus>(
+        streams.stream(0), streams.gpu_index(0), tmp_input_bits_b,
+        bits_per_block * num_radix_blocks, params.big_lwe_dimension,
+        size_tracker, allocate_gpu_memory);
+
+    tmp_mux_inputs = new CudaRadixCiphertextFFI;
+    create_zero_radix_ciphertext_async<Torus>(
+        streams.stream(0), streams.gpu_index(0), tmp_mux_inputs,
+        bits_per_block * num_radix_blocks, params.big_lwe_dimension,
+        size_tracker, allocate_gpu_memory);
+
+    auto mux_lut_f = [](Torus x) -> Torus {
+      // x is expected to be x = 0bcba
+      // where
+      // - c is the control bit
+      // - b the bit value returned if c is 1
+      // - a the bit value returned if c is 0
+      // (any bit above c is ignored)
+      x = x & 7;
+      auto control_bit = x >> 2;
+      auto previous_bit = (x & 2) >> 1;
+      auto current_bit = x & 1;
+
+      if (control_bit == 1)
+        return previous_bit;
+      else
+        return current_bit;
+    };
+
+    generate_device_accumulator<Torus>(
+        streams.stream(0), streams.gpu_index(0), mux_lut->get_lut(0, 0),
+        mux_lut->get_degree(0), mux_lut->get_max_degree(0),
+        params.glwe_dimension, params.polynomial_size, params.message_modulus,
+        params.carry_modulus, mux_lut_f, gpu_memory_allocated);
+    auto active_gpu_count_mux =
+        streams.active_gpu_subset(bits_per_block * num_radix_blocks);
+    mux_lut->broadcast_lut(active_gpu_count_mux);
+
+    auto cleaning_lut_f = [params](Torus x) -> Torus {
+      return x % params.message_modulus;
+    };
+    generate_device_accumulator<Torus>(
+        streams.stream(0), streams.gpu_index(0), cleaning_lut->get_lut(0, 0),
+        cleaning_lut->get_degree(0), cleaning_lut->get_max_degree(0),
+        params.glwe_dimension, params.polynomial_size, params.message_modulus,
+        params.carry_modulus, cleaning_lut_f, gpu_memory_allocated);
+    auto active_gpu_count_cleaning =
+        streams.active_gpu_subset(num_radix_blocks);
+    cleaning_lut->broadcast_lut(active_gpu_count_cleaning);
+  }
+
+  void release(CudaStreams streams) {
+    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
+                                   tmp_bits, gpu_memory_allocated);
+    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
+                                   tmp_shift_bits, gpu_memory_allocated);
+    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
+                                   tmp_rotated, gpu_memory_allocated);
+    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
+                                   tmp_input_bits_a, gpu_memory_allocated);
+    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
+                                   tmp_input_bits_b, gpu_memory_allocated);
+    release_radix_ciphertext_async(streams.stream(0), streams.gpu_index(0),
+                                   tmp_mux_inputs, gpu_memory_allocated);
+
+    bit_extract_luts->release(streams);
+    bit_extract_luts_with_offset_2->release(streams);
+    mux_lut->release(streams);
+    cleaning_lut->release(streams);
+
+    delete tmp_bits;
+    delete tmp_shift_bits;
+    delete tmp_rotated;
+    delete tmp_input_bits_a;
+    delete tmp_input_bits_b;
+    delete tmp_mux_inputs;
+    delete bit_extract_luts;
+    delete bit_extract_luts_with_offset_2;
+    delete mux_lut;
+    delete cleaning_lut;
+  }
+};
