+++ backends/tfhe-cuda-backend/cuda/src/integer/integer.cuh
@@ -4,13 +4,12 @@
 #include "crypto/keyswitch.cuh"
 #include "device.h"
 #include "helper_multi_gpu.h"
+#include "integer/comparison.h"
 #include "integer/integer_utilities.h"
 #include "integer/scalar_addition.cuh"
-#include "linear_algebra.h"
 #include "linearalgebra/addition.cuh"
 #include "linearalgebra/negation.cuh"
 #include "pbs/pbs_128_utilities.h"
-#include "pbs/programmable_bootstrap.h"
 #include "polynomial/functions.cuh"
 #include "utils/helper.cuh"
 #include "utils/helper_multi_gpu.cuh"
