+++ backends/tfhe-cuda-backend/cuda/src/integer/negation.cuh
@@ -10,13 +10,8 @@
 #include "device.h"
 #include "integer/integer.cuh"
 #include "integer/integer_utilities.h"
-#include "linear_algebra.h"
-#include "pbs/programmable_bootstrap.h"
-#include "utils/helper.cuh"
 #include "utils/kernel_dimensions.cuh"
-#include <fstream>
 #include <iostream>
-#include <omp.h>
 #include <sstream>
 #include <string>
 #include <vector>
@@ -109,63 +104,6 @@ __host__ void host_integer_radix_negation(
                       carry_modulus);
     zb = z / message_modulus;
   }
-}
-
-template <typename Torus>
-__host__ uint64_t scratch_cuda_integer_overflowing_sub_kb(
-    CudaStreams streams, int_overflowing_sub_memory<Torus> **mem_ptr,
-    uint32_t num_blocks, int_radix_params params, bool allocate_gpu_memory,
-    PBS_MS_REDUCTION_T noise_reduction_type) {
-
-  PUSH_RANGE("scratch overflowing sub")
-  uint64_t size_tracker = 0;
-  *mem_ptr = new int_overflowing_sub_memory<Torus>(
-      streams, params, num_blocks, allocate_gpu_memory, noise_reduction_type,
-      size_tracker);
-  POP_RANGE()
-  return size_tracker;
-}
-
-template <typename Torus>
-__host__ void host_integer_overflowing_sub(
-    CudaStreams streams, CudaRadixCiphertextFFI *output,
-    CudaRadixCiphertextFFI *input_left,
-    const CudaRadixCiphertextFFI *input_right,
-    CudaRadixCiphertextFFI *overflow_block,
-    const CudaRadixCiphertextFFI *input_borrow,
-    int_borrow_prop_memory<uint64_t> *mem_ptr, void *const *bsks,
-    Torus *const *ksks, uint32_t compute_overflow, uint32_t uses_input_borrow) {
-  PUSH_RANGE("overflowing sub")
-  if (output->num_radix_blocks != input_left->num_radix_blocks ||
-      output->num_radix_blocks != input_right->num_radix_blocks)
-    PANIC("Cuda error: lwe_array_in and output num radix blocks must be "
-          "the same")
-
-  if (output->lwe_dimension != input_left->lwe_dimension ||
-      output->lwe_dimension != input_right->lwe_dimension)
-    PANIC("Cuda error: lwe_array_in and output lwe_dimension must be "
-          "the same")
-
-  auto num_blocks = output->num_radix_blocks;
-  auto radix_params = mem_ptr->params;
-
-  // We need to recalculate the num_groups, because on the division the number
-  // of num_blocks changes
-  uint32_t block_modulus =
-      radix_params.message_modulus * radix_params.carry_modulus;
-  uint32_t num_bits_in_block = log2_int(block_modulus);
-  uint32_t grouping_size = num_bits_in_block;
-  uint32_t num_groups = (num_blocks + grouping_size - 1) / grouping_size;
-
-  host_unchecked_sub_with_correcting_term<Torus>(
-      streams.stream(0), streams.gpu_index(0), output, input_left, input_right,
-      num_blocks, radix_params.message_modulus, radix_params.carry_modulus);
-
-  host_single_borrow_propagate<Torus>(
-      streams, output, overflow_block, input_borrow,
-      (int_borrow_prop_memory<Torus> *)mem_ptr, bsks, (Torus **)(ksks),
-      num_groups, compute_overflow, uses_input_borrow);
-  POP_RANGE()
 }
 
 #endif
