+++ backends/tfhe-cuda-backend/cuda/src/integer/scalar_shifts.cuh
@@ -5,11 +5,9 @@
 #include "device.h"
 #include "integer.cuh"
 #include "integer/integer_utilities.h"
+#include "integer/scalar_shifts.h"
 #include "pbs/programmable_bootstrap_classic.cuh"
 #include "pbs/programmable_bootstrap_multibit.cuh"
-#include "types/complex/operations.cuh"
-#include "utils/helper.cuh"
-#include "utils/kernel_dimensions.cuh"
 
 template <typename Torus>
 __host__ uint64_t scratch_cuda_integer_radix_logical_scalar_shift_kb(
